Add CI workflows and extract ~2,500 lines from generator.py into modules
- Add GitHub Actions workflows for pytest and ruff linting
- Add ruff dev dependency and configuration in pyproject.toml
- Extract ConfigValidator, validate_credentials, validate_config_file
  to core/config_validation.py
- Extract CredentialLoader hierarchy, CredentialResolver, and credential
  helper functions to core/credentials.py
- Extract configure_cjapy, initialize_cja, _config_from_env to api/client.py
- Extract ParallelAPIFetcher to api/fetch.py
- Extract DataQualityChecker to api/quality.py
- Extract APIWorkerTuner to api/tuning.py
- Extract ValidationCache and SharedValidationCache to api/cache.py
- Extract JSONFormatter and setup_logging to core/logging.py
- Extract PerformanceTracker to core/perf.py
- Update test mock targets to match new module locations
- All 1,319 tests pass (1 skipped)
