+++ pyproject.toml
@@ -58,4 +58,35 @@ include = [
 dev = [
     "pytest>=8.3.4",
     "pytest-cov>=4.1.0",
+    "ruff>=0.9.0",
 ]
+
+[tool.ruff]
+target-version = "py314"
+line-length = 120
+src = ["src", "tests"]
+
+[tool.ruff.lint]
+select = [
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # pyflakes
+    "I",   # isort
+    "UP",  # pyupgrade
+    "B",   # flake8-bugbear
+    "SIM", # flake8-simplify
+    "RUF", # ruff-specific rules
+]
+ignore = [
+    "E501",   # line too long (handled by formatter)
+    "SIM108", # ternary operator (readability preference)
+    "B905",   # zip strict (Python 3.10+ only pattern)
+    "UP038",  # isinstance union type (readability preference)
+]
+
+[tool.ruff.lint.per-file-ignores]
+"src/cja_auto_sdr/generator.py" = ["E", "W", "F", "B", "SIM", "RUF", "UP"]
+"tests/**" = ["B", "SIM", "RUF012"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["cja_auto_sdr"]
