+++ src/cja_auto_sdr/api/quality.py
@@ -1,7 +1,442 @@
-"""Data quality validation (current implementation lives in generator)."""
+"""Data quality validation for CJA Auto SDR."""
 
-__all__ = ["DataQualityChecker"]
+import logging
+import threading
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from typing import Dict, List, Optional
 
-from cja_auto_sdr.core.lazy import make_getattr
+import pandas as pd
+from tqdm import tqdm
 
-__getattr__ = make_getattr(__name__, __all__, target_module="cja_auto_sdr.generator")
+from cja_auto_sdr.api.cache import ValidationCache
+from cja_auto_sdr.core.colors import _format_error_msg
+
+TQDM_BAR_FORMAT = "{l_bar}{bar}| {n_fmt}/{total_fmt} [{elapsed}]"
+
+
+class DataQualityChecker:
+    # Severity levels in priority order (highest to lowest) for proper sorting
+    SEVERITY_ORDER = ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW', 'INFO']
+
+    def __init__(self, logger: logging.Logger, validation_cache: Optional[ValidationCache] = None,
+                 quiet: bool = False):
+        self.issues = []
+        self.logger = logger
+        self.validation_cache = validation_cache  # Optional cache for performance
+        self._issues_lock = threading.Lock()  # Thread safety for parallel validation
+        self.quiet = quiet
+
+    def add_issue(self, severity: str, category: str, item_type: str,
+                  item_name: str, description: str, details: str = ""):
+        """Add a data quality issue to the tracker (thread-safe)"""
+        issue = {
+            'Severity': severity,
+            'Category': category,
+            'Type': item_type,
+            'Item Name': item_name,
+            'Issue': description,
+            'Details': details
+        }
+
+        # Thread-safe append operation
+        with self._issues_lock:
+            self.issues.append(issue)
+
+        # Conditional logging based on log level for performance
+        # Only log individual issues in DEBUG mode
+        if self.logger.isEnabledFor(logging.DEBUG):
+            self.logger.debug(f"DQ Issue [{severity}] - {item_type}: {description}")
+        elif severity in ['CRITICAL', 'HIGH'] and self.logger.isEnabledFor(logging.WARNING):
+            # In non-DEBUG modes, only log CRITICAL/HIGH severity issues
+            self.logger.warning(f"DQ Issue [{severity}] - {item_type}: {description}")
+
+    def check_duplicates(self, df: pd.DataFrame, item_type: str):
+        """Check for duplicate names in metrics or dimensions"""
+        try:
+            if df.empty:
+                self.logger.info(f"Skipping duplicate check for empty {item_type} dataframe")
+                return
+
+            if 'name' not in df.columns:
+                self.logger.warning(f"'name' column not found in {item_type}. Skipping duplicate check.")
+                return
+
+            duplicates = df['name'].value_counts()
+            duplicates = duplicates[duplicates > 1]
+
+            for name, count in duplicates.items():
+                self.add_issue(
+                    severity='HIGH',
+                    category='Duplicates',
+                    item_type=item_type,
+                    item_name=str(name),
+                    description=f'Duplicate name found {count} times',
+                    details=f'This {item_type.lower()} name appears {count} times in the data view'
+                )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking duplicates", item_type, e))
+
+    def check_required_fields(self, df: pd.DataFrame, item_type: str,
+                            required_fields: List[str]):
+        """Validate that required fields are present"""
+        try:
+            if df.empty:
+                self.logger.info(f"Skipping required fields check for empty {item_type} dataframe")
+                return
+
+            missing_fields = [field for field in required_fields if field not in df.columns]
+
+            if missing_fields:
+                self.add_issue(
+                    severity='CRITICAL',
+                    category='Missing Fields',
+                    item_type=item_type,
+                    item_name='N/A',
+                    description=f'Required fields missing from API response',
+                    details=f'Missing fields: {", ".join(missing_fields)}'
+                )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking required fields", item_type, e))
+
+    def check_null_values(self, df: pd.DataFrame, item_type: str,
+                         critical_fields: List[str]):
+        """Check for null values in critical fields"""
+        try:
+            if df.empty:
+                self.logger.info(f"Skipping null value check for empty {item_type} dataframe")
+                return
+
+            for field in critical_fields:
+                if field in df.columns:
+                    null_count = df[field].isna().sum()
+                    if null_count > 0:
+                        null_items = df[df[field].isna()]['name'].tolist() if 'name' in df.columns else []
+                        self.add_issue(
+                            severity='MEDIUM',
+                            category='Null Values',
+                            item_type=item_type,
+                            item_name=', '.join(str(x) for x in null_items),
+                            description=f'Null values in "{field}" field',
+                            details=f'{null_count} item(s) missing {field}. Items: {", ".join(str(x) for x in null_items)}'
+                        )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking null values", item_type, e))
+
+    def check_missing_descriptions(self, df: pd.DataFrame, item_type: str):
+        """Check for items without descriptions"""
+        try:
+            if df.empty:
+                self.logger.info(f"Skipping description check for empty {item_type} dataframe")
+                return
+
+            if 'description' not in df.columns:
+                self.logger.info(f"'description' column not found in {item_type}")
+                return
+
+            missing_desc = df[df['description'].isna() | (df['description'] == '')]
+
+            if len(missing_desc) > 0:
+                item_names = missing_desc['name'].tolist() if 'name' in missing_desc.columns else []
+                self.add_issue(
+                    severity='LOW',
+                    category='Missing Descriptions',
+                    item_type=item_type,
+                    item_name=f'{len(missing_desc)} items',
+                    description=f'{len(missing_desc)} items without descriptions',
+                    details=f'Items: {", ".join(str(x) for x in item_names)}'
+                )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking descriptions", item_type, e))
+
+    def check_empty_dataframe(self, df: pd.DataFrame, item_type: str):
+        """Check if dataframe is empty"""
+        try:
+            if df.empty:
+                self.add_issue(
+                    severity='CRITICAL',
+                    category='Empty Data',
+                    item_type=item_type,
+                    item_name='N/A',
+                    description=f'No {item_type.lower()} found in data view',
+                    details=f'The API returned an empty dataset for {item_type.lower()}'
+                )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking if dataframe is empty", item_type, e))
+
+    def check_id_validity(self, df: pd.DataFrame, item_type: str):
+        """Check for missing or invalid IDs"""
+        try:
+            if df.empty:
+                self.logger.info(f"Skipping ID validity check for empty {item_type} dataframe")
+                return
+
+            if 'id' not in df.columns:
+                self.logger.warning(f"'id' column not found in {item_type}")
+                return
+
+            missing_ids = df[df['id'].isna() | (df['id'] == '')]
+            if len(missing_ids) > 0:
+                self.add_issue(
+                    severity='HIGH',
+                    category='Invalid IDs',
+                    item_type=item_type,
+                    item_name=f'{len(missing_ids)} items',
+                    description=f'{len(missing_ids)} items with missing or invalid IDs',
+                    details='Items without valid IDs may cause issues in reporting'
+                )
+        except Exception as e:
+            self.logger.error(_format_error_msg("checking ID validity", item_type, e))
+
+    def check_all_quality_issues_optimized(self, df: pd.DataFrame, item_type: str,
+                                           required_fields: List[str],
+                                           critical_fields: List[str]):
+        """
+        Optimized single-pass validation combining all checks
+
+        PERFORMANCE OPTIMIZATIONS:
+        - 40-55% faster than sequential individual checks
+        - Reduces DataFrame scans from 6 to 1
+        - Uses vectorized pandas operations
+        - Early exit on critical errors (5-10% additional improvement)
+        - Validation caching (50-90% improvement on cache hits)
+        - Better CPU cache utilization
+        """
+        try:
+            # Check cache first (before any processing)
+            cache_key = None
+            if self.validation_cache is not None:
+                cached_issues, cache_key = self.validation_cache.get(
+                    df, item_type, required_fields, critical_fields
+                )
+                if cached_issues is not None:
+                    with self._issues_lock:
+                        self.issues.extend(cached_issues)
+                    self.logger.debug(f"Using cached validation results for {item_type}")
+                    return
+
+            issues_start_index = len(self.issues)
+
+            # Check 1: Empty DataFrame (quick exit)
+            if df.empty:
+                self.add_issue(
+                    severity='CRITICAL',
+                    category='Empty Data',
+                    item_type=item_type,
+                    item_name='N/A',
+                    description=f'No {item_type.lower()} found in data view',
+                    details=f'The API returned an empty dataset for {item_type.lower()}'
+                )
+                if self.validation_cache is not None:
+                    new_issues = self.issues[issues_start_index:]
+                    self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
+                return
+
+            # Check 2: Required fields validation
+            missing_fields = [field for field in required_fields if field not in df.columns]
+            if missing_fields:
+                self.add_issue(
+                    severity='CRITICAL',
+                    category='Missing Fields',
+                    item_type=item_type,
+                    item_name='N/A',
+                    description='Required fields missing from API response',
+                    details=f'Missing fields: {", ".join(missing_fields)}'
+                )
+                if self.validation_cache is not None:
+                    new_issues = self.issues[issues_start_index:]
+                    self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
+                return
+
+            # Check 3: Vectorized duplicate detection
+            if 'name' in df.columns:
+                duplicates = df['name'].value_counts()
+                duplicates = duplicates[duplicates > 1]
+                for name, count in duplicates.items():
+                    self.add_issue(
+                        severity='HIGH',
+                        category='Duplicates',
+                        item_type=item_type,
+                        item_name=str(name),
+                        description=f'Duplicate name found {count} times',
+                        details=f'This {item_type.lower()} name appears {count} times in the data view'
+                    )
+
+            # Check 4: Vectorized null value checks
+            available_critical_fields = [f for f in critical_fields if f in df.columns]
+            if available_critical_fields:
+                null_counts = df[available_critical_fields].isna().sum()
+                for field, null_count in null_counts[null_counts > 0].items():
+                    null_items = df[df[field].isna()]['name'].tolist() if 'name' in df.columns else []
+                    self.add_issue(
+                        severity='MEDIUM',
+                        category='Null Values',
+                        item_type=item_type,
+                        item_name=', '.join(str(x) for x in null_items),
+                        description=f'Null values in "{field}" field',
+                        details=f'{null_count} item(s) missing {field}. Items: {", ".join(str(x) for x in null_items)}'
+                    )
+
+            # Check 5: Vectorized missing descriptions check
+            if 'description' in df.columns:
+                missing_desc = df[df['description'].isna() | (df['description'] == '')]
+                if len(missing_desc) > 0:
+                    item_names = missing_desc['name'].tolist() if 'name' in missing_desc.columns else []
+                    self.add_issue(
+                        severity='LOW',
+                        category='Missing Descriptions',
+                        item_type=item_type,
+                        item_name=f'{len(missing_desc)} items',
+                        description=f'{len(missing_desc)} items without descriptions',
+                        details=f'Items: {", ".join(str(x) for x in item_names)}'
+                    )
+
+            # Check 6: Vectorized ID validity check
+            if 'id' in df.columns:
+                missing_ids = df[df['id'].isna() | (df['id'] == '')]
+                if len(missing_ids) > 0:
+                    self.add_issue(
+                        severity='HIGH',
+                        category='Invalid IDs',
+                        item_type=item_type,
+                        item_name=f'{len(missing_ids)} items',
+                        description=f'{len(missing_ids)} items with missing or invalid IDs',
+                        details='Items without valid IDs may cause issues in reporting'
+                    )
+
+            if self.logger.isEnabledFor(logging.DEBUG):
+                self.logger.debug(f"Optimized validation complete for {item_type}: {len(df)} items checked")
+
+            if self.validation_cache is not None:
+                new_issues = self.issues[issues_start_index:]
+                self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
+
+        except Exception as e:
+            self.logger.error(_format_error_msg("in optimized validation", item_type, e))
+            self.logger.exception("Full error details:")
+
+    def check_all_parallel(self,
+                          metrics_df: pd.DataFrame,
+                          dimensions_df: pd.DataFrame,
+                          metrics_required_fields: List[str],
+                          dimensions_required_fields: List[str],
+                          critical_fields: List[str],
+                          max_workers: int = 2):
+        """
+        Run validation checks in parallel for metrics and dimensions
+        """
+        try:
+            self.logger.info("Starting parallel validation (metrics and dimensions)")
+
+            tasks = {
+                'metrics': lambda: self.check_all_quality_issues_optimized(
+                    metrics_df, 'Metrics', metrics_required_fields, critical_fields
+                ),
+                'dimensions': lambda: self.check_all_quality_issues_optimized(
+                    dimensions_df, 'Dimensions', dimensions_required_fields, critical_fields
+                )
+            }
+
+            with ThreadPoolExecutor(max_workers=max_workers) as executor:
+                future_to_name = {
+                    executor.submit(task): name
+                    for name, task in tasks.items()
+                }
+
+                with tqdm(
+                    total=len(tasks),
+                    desc="Validating data",
+                    unit="check",
+                    bar_format=TQDM_BAR_FORMAT,
+                    leave=False,
+                    disable=self.quiet
+                ) as pbar:
+                    for future in as_completed(future_to_name):
+                        task_name = future_to_name[future]
+                        try:
+                            future.result()
+                            pbar.set_postfix_str(f"\u2713 {task_name}", refresh=True)
+                            self.logger.debug(f"\u2713 {task_name.capitalize()} validation completed")
+                        except Exception as e:
+                            pbar.set_postfix_str(f"\u2717 {task_name}", refresh=True)
+                            self.logger.error(f"\u2717 {task_name.capitalize()} validation failed: {e}")
+                            self.logger.exception("Full error details:")
+                        pbar.update(1)
+
+            self.logger.info(f"Parallel validation complete. Found {len(self.issues)} issue(s)")
+
+        except Exception as e:
+            self.logger.error(_format_error_msg("in parallel validation", error=e))
+            self.logger.exception("Full error details:")
+            raise
+
+    def get_issues_dataframe(self, max_issues: int = 0) -> pd.DataFrame:
+        """Return all issues as a DataFrame sorted by severity (CRITICAL first)
+
+        Args:
+            max_issues: Maximum number of issues to return (0 = all issues)
+        """
+        try:
+            if not self.issues:
+                self.logger.info("No data quality issues found")
+                return pd.DataFrame({
+                    'Severity': ['INFO'],
+                    'Category': ['Data Quality'],
+                    'Type': ['All'],
+                    'Item Name': ['N/A'],
+                    'Issue': ['No data quality issues detected'],
+                    'Details': ['All validation checks passed successfully']
+                })
+
+            df = pd.DataFrame(self.issues)
+
+            # Use CategoricalDtype for proper severity ordering (CRITICAL > HIGH > MEDIUM > LOW > INFO)
+            severity_dtype = pd.CategoricalDtype(categories=self.SEVERITY_ORDER, ordered=True)
+            df['Severity'] = df['Severity'].astype(severity_dtype)
+
+            # Reorder columns: Severity first for better readability
+            preferred_order = ['Severity', 'Category', 'Type', 'Item Name', 'Issue', 'Details']
+            existing_cols = [col for col in preferred_order if col in df.columns]
+            other_cols = [col for col in df.columns if col not in preferred_order]
+            df = df[existing_cols + other_cols]
+
+            # Sort by severity then by Category alphabetically
+            df = df.sort_values(
+                by=['Severity', 'Category'],
+                ascending=[True, True]
+            )
+
+            # Limit to top N issues if max_issues > 0
+            if max_issues > 0 and len(df) > max_issues:
+                self.logger.info(f"Limiting data quality issues to top {max_issues} (of {len(df)} total)")
+                df = df.head(max_issues)
+
+            return df
+        except Exception as e:
+            self.logger.error(_format_error_msg("creating issues dataframe", error=e))
+            return pd.DataFrame({
+                'Severity': ['ERROR'],
+                'Category': ['System'],
+                'Type': ['Processing'],
+                'Item Name': ['N/A'],
+                'Issue': ['Error generating data quality report'],
+                'Details': [str(e)]
+            })
+
+    def log_summary(self):
+        """Log aggregated summary of data quality issues for performance"""
+        if not self.issues:
+            self.logger.info("\u2713 No data quality issues found")
+            return
+
+        # Aggregate by severity
+        severity_counts: Dict[str, int] = {}
+        for issue in self.issues:
+            sev = issue['Severity']
+            severity_counts[sev] = severity_counts.get(sev, 0) + 1
+
+        self.logger.info(f"Data quality validation complete: {len(self.issues)} issue(s) found")
+
+        if self.logger.isEnabledFor(logging.INFO):
+            for sev in ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW']:
+                if sev in severity_counts:
+                    self.logger.info(f"  {sev}: {severity_counts[sev]}")
