+++ src/cja_auto_sdr/generator.py
@@ -350,855 +350,16 @@ from cja_auto_sdr.diff.snapshot import SnapshotManager, parse_retention_period
 # ==================== LOGGING SETUP ====================
 
 
-class JSONFormatter(logging.Formatter):
-    """
-    JSON formatter for structured logging output.
+# ==================== LOGGING (moved to core/logging.py) ====================
+from cja_auto_sdr.core.logging import JSONFormatter, setup_logging  # noqa: F811
 
-    Produces JSON lines suitable for log aggregation systems (Splunk, ELK, CloudWatch).
-    Each log record is a single JSON object on one line.
-    """
+# ==================== PERFORMANCE TRACKING (moved to core/perf.py) ====================
+from cja_auto_sdr.core.perf import PerformanceTracker  # noqa: F811
 
-    def format(self, record: logging.LogRecord) -> str:
-        """Format log record as JSON."""
-        log_entry = {
-            "timestamp": datetime.fromtimestamp(record.created).isoformat(),
-            "level": record.levelname,
-            "logger": record.name,
-            "message": record.getMessage(),
-            "module": record.module,
-            "function": record.funcName,
-            "line": record.lineno,
-        }
+# ==================== API WORKER TUNER (moved to api/tuning.py) ====================
+from cja_auto_sdr.api.tuning import APIWorkerTuner  # noqa: F811
 
-        # Add exception info if present
-        if record.exc_info:
-            log_entry["exception"] = self.formatException(record.exc_info)
-
-        # Add any extra fields passed to the logger
-        if hasattr(record, 'extra_fields'):
-            log_entry.update(record.extra_fields)
-
-        return json.dumps(log_entry, default=str)
-
-
-# Module-level tracking to prevent duplicate logger initialization
-_logging_initialized = False
-_current_log_file = None
-_atexit_registered = False
-
-def setup_logging(
-    data_view_id: Optional[str] = None,
-    batch_mode: bool = False,
-    log_level: Optional[str] = None,
-    log_format: str = "text"
-) -> logging.Logger:
-    """Setup logging to both file and console.
-
-    Args:
-        data_view_id: Data view ID for log file naming
-        batch_mode: Whether running in batch mode
-        log_level: Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
-        log_format: Output format - "text" (default) or "json" for structured logging
-
-    Returns:
-        Configured logger instance
-
-    Priority: 1) Passed parameter, 2) Environment variable LOG_LEVEL, 3) Default INFO
-    """
-    global _logging_initialized, _current_log_file, _atexit_registered
-
-    # Register atexit handler once to ensure logs are flushed on exit
-    if not _atexit_registered:
-        atexit.register(logging.shutdown)
-        _atexit_registered = True
-
-    # Create logs directory if it doesn't exist
-    log_dir = Path("logs")
-    try:
-        log_dir.mkdir(exist_ok=True)
-    except PermissionError:
-        print(f"Warning: Cannot create logs directory (permission denied). Logging to console only.", file=sys.stderr)
-        log_dir = None
-    except OSError as e:
-        print(f"Warning: Cannot create logs directory: {e}. Logging to console only.", file=sys.stderr)
-        log_dir = None
-
-    # Create log filename with timestamp
-    timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-
-    if log_dir is not None:
-        if batch_mode:
-            log_file = log_dir / f"SDR_Batch_Generation_{timestamp}.log"
-        else:
-            log_file = log_dir / f"SDR_Generation_{data_view_id}_{timestamp}.log"
-    else:
-        log_file = None
-
-    # Determine log level with priority: parameter > env var > default
-    if log_level is None:
-        log_level = os.environ.get('LOG_LEVEL', 'INFO')
-
-    # Validate log level
-    valid_levels = ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']
-    if log_level.upper() not in valid_levels:
-        print(f"Warning: Invalid log level '{log_level}', using INFO", file=sys.stderr)
-        log_level = 'INFO'
-
-    # Get numeric log level
-    numeric_level = getattr(logging, log_level.upper(), logging.INFO)
-
-    # Clear any existing handlers from root logger
-    for handler in logging.root.handlers[:]:
-        handler.close()
-        logging.root.removeHandler(handler)
-
-    # Configure logging handlers
-    handlers = [logging.StreamHandler(sys.stdout)]
-    if log_file is not None:
-        # Use RotatingFileHandler to prevent unbounded log growth
-        handlers.append(RotatingFileHandler(
-            log_file,
-            maxBytes=LOG_FILE_MAX_BYTES,
-            backupCount=LOG_FILE_BACKUP_COUNT
-        ))
-
-    # Select formatter based on log_format
-    if log_format.lower() == "json":
-        formatter = JSONFormatter()
-    else:
-        formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-
-    # Apply formatter and level to all handlers, then add to root logger
-    for handler in handlers:
-        handler.setFormatter(formatter)
-        handler.setLevel(numeric_level)
-        logging.root.addHandler(handler)
-
-    # Set root logger level explicitly
-    logging.root.setLevel(numeric_level)
-
-    # Get the module logger
-    logger = logging.getLogger(__name__)
-    # Ensure it propagates to root and doesn't have its own restrictive level
-    logger.propagate = True
-    logger.setLevel(logging.NOTSET)
-
-    # Track initialization state to prevent duplicates
-    _logging_initialized = True
-    _current_log_file = log_file
-
-    if log_file is not None:
-        logger.info(f"Logging initialized. Log file: {log_file}")
-    else:
-        logger.info("Logging initialized. Console output only.")
-
-    # Flush handlers to ensure log file is not empty even on early exit
-    for handler in logging.root.handlers:
-        handler.flush()
-
-    return logger
-
-# ==================== PERFORMANCE TRACKING ====================
-
-class PerformanceTracker:
-    """Track execution time for operations"""
-    MAX_METRICS = 500
-
-    def __init__(self, logger: logging.Logger):
-        self.metrics = {}
-        self.logger = logger
-        self.start_times = {}
-
-    def start(self, operation_name: str):
-        """Start timing an operation"""
-        self.start_times[operation_name] = time.time()
-
-    def end(self, operation_name: str):
-        """End timing an operation"""
-        if operation_name in self.start_times:
-            duration = time.time() - self.start_times[operation_name]
-            if len(self.metrics) >= self.MAX_METRICS:
-                # Drop oldest entry to prevent unbounded growth
-                oldest = next(iter(self.metrics))
-                del self.metrics[oldest]
-            self.metrics[operation_name] = duration
-
-            # Log individual operations only in DEBUG mode for performance
-            if self.logger.isEnabledFor(logging.DEBUG):
-                self.logger.debug(f"⏱️  {operation_name} completed in {duration:.2f}s")
-
-            del self.start_times[operation_name]
-
-    def get_summary(self) -> str:
-        """Generate performance summary"""
-        if not self.metrics:
-            return "No performance metrics collected"
-
-        total = sum(self.metrics.values())
-        lines = ["", "=" * BANNER_WIDTH, "PERFORMANCE SUMMARY", "=" * BANNER_WIDTH]
-
-        for operation, duration in sorted(self.metrics.items(), key=lambda x: x[1], reverse=True):
-            percentage = (duration / total) * 100 if total > 0 else 0
-            lines.append(f"{operation:35s}: {duration:6.2f}s ({percentage:5.1f}%)")
-
-        lines.extend(["=" * BANNER_WIDTH, f"{'Total Execution Time':35s}: {total:6.2f}s", "=" * BANNER_WIDTH])
-        return "\n".join(lines)
-
-    def add_cache_statistics(self, cache):
-        """Add cache statistics to performance metrics"""
-        stats = cache.get_statistics()
-
-        if stats['total_requests'] > 0:
-            self.logger.info("")
-            self.logger.info("=" * BANNER_WIDTH)
-            self.logger.info("VALIDATION CACHE STATISTICS")
-            self.logger.info("=" * BANNER_WIDTH)
-            self.logger.info(f"Cache Hits:        {stats['hits']}")
-            self.logger.info(f"Cache Misses:      {stats['misses']}")
-            self.logger.info(f"Hit Rate:          {stats['hit_rate']:.1f}%")
-            self.logger.info(f"Cache Size:        {stats['size']}/{stats['max_size']}")
-            self.logger.info(f"Evictions:         {stats['evictions']}")
-
-            if stats['hits'] > 0:
-                # Assume average validation takes 50ms, cache lookup takes 1ms
-                time_saved = stats['hits'] * 0.049  # 49ms saved per hit
-                self.logger.info(f"Estimated Time Saved: {time_saved:.2f}s")
-
-            self.logger.info("=" * BANNER_WIDTH)
-
-
-# ==================== API WORKER TUNER ====================
-
-class APIWorkerTuner:
-    """
-    Thread-safe auto-tuner for API worker pool size.
-
-    Dynamically adjusts the number of concurrent API workers based on
-    response time measurements. Scales up when responses are fast (API
-    has capacity) and scales down when responses are slow (API is stressed).
-
-    Algorithm:
-    - Maintains a rolling window of response times
-    - After sample_window measurements, calculates average
-    - If avg < scale_up_threshold_ms: increase workers (if below max)
-    - If avg > scale_down_threshold_ms: decrease workers (if above min)
-    - Enforces cooldown period between adjustments
-
-    Thread Safety:
-    - Uses threading.Lock for all state updates
-    - Safe for use with ThreadPoolExecutor
-
-    Example:
-        tuner = APIWorkerTuner(config=APITuningConfig(min_workers=1, max_workers=10))
-
-        # After each API call
-        new_count = tuner.record_response_time(duration_ms)
-        if new_count is not None:
-            executor.resize(new_count)  # Adjust pool size
-    """
-
-    def __init__(self, config: APITuningConfig = None, initial_workers: int = 3,
-                 logger: logging.Logger = None):
-        """
-        Initialize the API worker tuner.
-
-        Args:
-            config: Tuning configuration (uses defaults if not provided)
-            initial_workers: Starting number of workers
-            logger: Logger instance for tuning decisions
-        """
-        self.config = config or APITuningConfig()
-        self.logger = logger or logging.getLogger(__name__)
-
-        # Current worker count
-        self._current_workers = max(
-            self.config.min_workers,
-            min(initial_workers, self.config.max_workers)
-        )
-
-        # Rolling window of response times
-        self._response_times: List[float] = []
-        self._last_adjustment_time = 0.0
-
-        # Thread safety
-        self._lock = threading.Lock()
-
-        # Statistics
-        self._total_requests = 0
-        self._scale_ups = 0
-        self._scale_downs = 0
-        self._total_response_time_ms = 0.0
-
-    @property
-    def current_workers(self) -> int:
-        """Get current worker count (thread-safe)."""
-        with self._lock:
-            return self._current_workers
-
-    def record_response_time(self, duration_ms: float) -> Optional[int]:
-        """
-        Record a response time and potentially adjust worker count.
-
-        Args:
-            duration_ms: Response time in milliseconds
-
-        Returns:
-            New worker count if adjusted, None if no change needed
-
-        Note:
-            Call this after each API request completes (success or failure).
-        """
-        with self._lock:
-            self._total_requests += 1
-            self._total_response_time_ms += duration_ms
-            self._response_times.append(duration_ms)
-
-            # Keep only the most recent samples
-            if len(self._response_times) > self.config.sample_window:
-                self._response_times = self._response_times[-self.config.sample_window:]
-
-            # Only adjust after we have enough samples
-            if len(self._response_times) < self.config.sample_window:
-                return None
-
-            # Check cooldown period
-            now = time.time()
-            if now - self._last_adjustment_time < self.config.cooldown_seconds:
-                return None
-
-            # Calculate average response time
-            avg_time = sum(self._response_times) / len(self._response_times)
-
-            # Determine if adjustment is needed
-            new_workers = None
-
-            if avg_time < self.config.scale_up_threshold_ms:
-                # Fast responses - try to add workers
-                if self._current_workers < self.config.max_workers:
-                    new_workers = self._current_workers + 1
-                    self._scale_ups += 1
-                    self.logger.info(
-                        f"API tuner: scaling UP {self._current_workers} → {new_workers} workers "
-                        f"(avg response: {avg_time:.0f}ms < {self.config.scale_up_threshold_ms}ms threshold)"
-                    )
-
-            elif avg_time > self.config.scale_down_threshold_ms:
-                # Slow responses - try to reduce workers
-                if self._current_workers > self.config.min_workers:
-                    new_workers = self._current_workers - 1
-                    self._scale_downs += 1
-                    self.logger.info(
-                        f"API tuner: scaling DOWN {self._current_workers} → {new_workers} workers "
-                        f"(avg response: {avg_time:.0f}ms > {self.config.scale_down_threshold_ms}ms threshold)"
-                    )
-
-            if new_workers is not None:
-                self._current_workers = new_workers
-                self._last_adjustment_time = now
-                self._response_times.clear()  # Reset window after adjustment
-                return new_workers
-
-            return None
-
-    def get_statistics(self) -> Dict[str, Any]:
-        """
-        Get tuner statistics.
-
-        Returns:
-            Dict with worker counts, adjustment history, and performance metrics
-        """
-        with self._lock:
-            avg_response = (
-                self._total_response_time_ms / self._total_requests
-                if self._total_requests > 0 else 0.0
-            )
-
-            return {
-                'current_workers': self._current_workers,
-                'min_workers': self.config.min_workers,
-                'max_workers': self.config.max_workers,
-                'total_requests': self._total_requests,
-                'scale_ups': self._scale_ups,
-                'scale_downs': self._scale_downs,
-                'average_response_ms': avg_response,
-                'sample_window_size': len(self._response_times),
-            }
-
-    def reset(self, workers: int = None) -> None:
-        """
-        Reset the tuner state.
-
-        Args:
-            workers: Reset to specific worker count (default: keep current)
-        """
-        with self._lock:
-            if workers is not None:
-                self._current_workers = max(
-                    self.config.min_workers,
-                    min(workers, self.config.max_workers)
-                )
-            self._response_times.clear()
-            self._last_adjustment_time = 0.0
-            self.logger.debug(f"API tuner reset (workers: {self._current_workers})")
-
-
-# ==================== VALIDATION CACHE ====================
-
-class ValidationCache:
-    """
-    Thread-safe LRU cache for data quality validation results
-
-    Caches validation results based on DataFrame content hash and configuration.
-    Uses LRU eviction policy to prevent unbounded memory growth.
-
-    Performance Impact:
-    - Cache hits: 50-90% faster than running validation
-    - Cache misses: ~1-2% overhead for hashing (negligible)
-    - Memory: ~1-5MB per 1000 cached entries
-
-    Thread Safety:
-    - Uses threading.Lock for all cache operations
-    - Safe for use with parallel validation (check_all_parallel)
-    """
-
-    def __init__(self, max_size: int = 1000, ttl_seconds: int = 3600, logger: logging.Logger = None):
-        """
-        Initialize validation cache
-
-        Args:
-            max_size: Maximum number of cached entries, >= 1 (default: 1000)
-            ttl_seconds: Time-to-live for cache entries in seconds, >= 1 (default: 3600 = 1 hour)
-            logger: Logger instance for cache statistics (default: module logger)
-        """
-        self.max_size = max_size
-        self.ttl_seconds = ttl_seconds
-        self.logger = logger or logging.getLogger(__name__)
-
-        # Cache storage: key -> (issues_list, timestamp)
-        self._cache: Dict[str, Tuple[List[Dict], float]] = {}
-
-        # LRU tracking: key -> last_access_time
-        self._access_times: Dict[str, float] = {}
-
-        # Thread safety
-        self._lock = threading.Lock()
-
-        # Statistics
-        self._hits = 0
-        self._misses = 0
-        self._evictions = 0
-
-        self.logger.debug(f"ValidationCache initialized: max_size={max_size}, ttl={ttl_seconds}s")
-
-    def _generate_cache_key(self, df: pd.DataFrame, item_type: str,
-                           required_fields: List[str], critical_fields: List[str]) -> str:
-        """
-        Generate cache key from DataFrame content and configuration
-
-        Strategy:
-        - Uses pandas.util.hash_pandas_object for efficient DataFrame hashing
-        - Combines DataFrame hash with configuration parameters
-        - Returns consistent hash for identical inputs
-
-        Args:
-            df: DataFrame to hash
-            item_type: 'Metrics' or 'Dimensions'
-            required_fields: List of required field names
-            critical_fields: List of critical field names
-
-        Returns:
-            Cache key string in format: "{item_type}:{df_hash}:{config_hash}"
-        """
-        try:
-            # Hash DataFrame content using pandas built-in function
-            # This is much faster than manual iteration (1-2ms vs 10-50ms for 1000 rows)
-
-            # Hash DataFrame structure and content
-            df_hash = pd.util.hash_pandas_object(df, index=False).sum()
-
-            # Hash configuration (required_fields + critical_fields)
-            config_str = f"{sorted(required_fields)}:{sorted(critical_fields)}"
-            config_hash = hashlib.md5(config_str.encode()).hexdigest()[:8]
-
-            # Combine into cache key
-            cache_key = f"{item_type}:{df_hash}:{config_hash}"
-
-            return cache_key
-
-        except Exception as e:
-            self.logger.warning(f"Error generating cache key: {e}. Cache disabled for this call.")
-            # Return unique key to force cache miss
-            return f"error:{time.time()}"
-
-    def get(self, df: pd.DataFrame, item_type: str,
-           required_fields: List[str], critical_fields: List[str]) -> Tuple[Optional[List[Dict]], str]:
-        """
-        Retrieve cached validation results if available
-
-        Returns:
-            Tuple of (issues list or None, cache_key).
-            The cache_key can be passed to put() to avoid recomputing the hash.
-        """
-        cache_key = self._generate_cache_key(df, item_type, required_fields, critical_fields)
-
-        # Check debug logging once outside the lock to avoid repeated checks
-        debug_enabled = self.logger.isEnabledFor(logging.DEBUG)
-
-        with self._lock:
-            if cache_key not in self._cache:
-                self._misses += 1
-                if debug_enabled:
-                    self.logger.debug(f"Cache MISS: {item_type} (key: {cache_key[:20]}...)")
-                return None, cache_key
-
-            cached_issues, timestamp = self._cache[cache_key]
-
-            # Check TTL expiration
-            age = time.time() - timestamp
-            if age > self.ttl_seconds:
-                if debug_enabled:
-                    self.logger.debug(f"Cache EXPIRED: {item_type} (age: {age:.1f}s)")
-                del self._cache[cache_key]
-                del self._access_times[cache_key]
-                self._misses += 1
-                return None, cache_key
-
-            # Cache hit - update access time
-            self._access_times[cache_key] = time.time()
-            self._hits += 1
-            if debug_enabled:
-                self.logger.debug(f"Cache HIT: {item_type} ({len(cached_issues)} issues)")
-
-            # Return deep copy to prevent mutation of cached data
-            return [issue.copy() for issue in cached_issues], cache_key
-
-    def put(self, df: pd.DataFrame, item_type: str,
-           required_fields: List[str], critical_fields: List[str],
-           issues: List[Dict], cache_key: str = None):
-        """
-        Store validation results in cache
-
-        Implements LRU eviction when cache is full.
-
-        Args:
-            cache_key: Optional pre-computed cache key from get() to avoid rehashing
-        """
-        if cache_key is None:
-            cache_key = self._generate_cache_key(df, item_type, required_fields, critical_fields)
-
-        # Check debug logging once to avoid repeated checks in hot path
-        debug_enabled = self.logger.isEnabledFor(logging.DEBUG)
-
-        with self._lock:
-            # Evict oldest entry if cache is full
-            if len(self._cache) >= self.max_size and cache_key not in self._cache:
-                self._evict_lru(debug_enabled)
-
-            # Store issues with timestamp
-            # Deep copy to prevent external mutation
-            self._cache[cache_key] = ([issue.copy() for issue in issues], time.time())
-            self._access_times[cache_key] = time.time()
-
-            if debug_enabled:
-                self.logger.debug(f"Cache STORE: {item_type} ({len(issues)} issues)")
-
-    def _evict_lru(self, debug_enabled: bool = False):
-        """Evict least recently used cache entry.
-
-        Args:
-            debug_enabled: Whether debug logging is enabled (avoids repeated checks)
-        """
-        if not self._access_times:
-            return
-
-        # Find least recently used key
-        lru_key = min(self._access_times.items(), key=lambda x: x[1])[0]
-
-        # Remove from cache
-        del self._cache[lru_key]
-        del self._access_times[lru_key]
-        self._evictions += 1
-
-        if debug_enabled:
-            self.logger.debug(f"Cache EVICT: LRU entry removed (total evictions: {self._evictions})")
-
-    def get_statistics(self) -> Dict[str, Any]:
-        """
-        Get cache performance statistics
-
-        Returns:
-            Dict with hits, misses, hit_rate, size, evictions
-        """
-        with self._lock:
-            total_requests = self._hits + self._misses
-            hit_rate = (self._hits / total_requests * 100) if total_requests > 0 else 0
-
-            return {
-                'hits': self._hits,
-                'misses': self._misses,
-                'hit_rate': hit_rate,
-                'size': len(self._cache),
-                'max_size': self.max_size,
-                'evictions': self._evictions,
-                'total_requests': total_requests
-            }
-
-    def clear(self):
-        """Clear all cache entries (useful for testing)"""
-        with self._lock:
-            self._cache.clear()
-            self._access_times.clear()
-            self.logger.debug("Cache cleared")
-
-    def log_statistics(self):
-        """
-        Log cache statistics in a user-friendly format.
-
-        Logs hit rate, total requests, cache size, and estimated time savings.
-        Only logs if there have been cache requests.
-        """
-        stats = self.get_statistics()
-        if stats['total_requests'] == 0:
-            self.logger.debug("Cache statistics: No requests recorded")
-            return
-
-        # Estimate time saved (average validation ~50ms, cache lookup ~1ms)
-        estimated_time_saved = stats['hits'] * 0.049  # 49ms saved per hit
-
-        self.logger.info(f"Cache Statistics: {stats['hits']}/{stats['total_requests']} hits "
-                        f"({stats['hit_rate']:.1f}% hit rate)")
-        self.logger.info(f"  - Cache size: {stats['size']}/{stats['max_size']} entries")
-        if stats['evictions'] > 0:
-            self.logger.info(f"  - Evictions: {stats['evictions']}")
-        if estimated_time_saved > 0.1:
-            self.logger.info(f"  - Estimated time saved: {estimated_time_saved:.2f}s")
-
-
-# ==================== SHARED VALIDATION CACHE ====================
-
-class SharedValidationCache:
-    """
-    Process-safe shared cache for validation results across batch workers.
-
-    Uses multiprocessing.Manager to share cache state across multiple
-    processes in batch mode. This allows validation results to be reused
-    across different data views being processed in parallel.
-
-    API Compatibility:
-    - Implements the same interface as ValidationCache for drop-in replacement
-    - get() and put() methods accept DataFrame objects like ValidationCache
-
-    Key Differences from ValidationCache:
-    - Uses Manager().dict() instead of regular dict (process-safe)
-    - Manager-level locking instead of threading.Lock
-    - Requires explicit shutdown() to cleanup Manager resources
-    - Slightly higher overhead but enables cross-process sharing
-
-    Usage:
-        # In BatchProcessor
-        shared_cache = SharedValidationCache(max_size=1000)
-
-        # Pass to workers (cache reference is pickle-safe via Manager)
-        worker_args = (data_view_id, ..., shared_cache)
-
-        # In worker process - same API as ValidationCache
-        cached, key = shared_cache.get(df, item_type, req_fields, crit_fields)
-        if cached is None:
-            result = run_validation(...)
-            shared_cache.put(df, item_type, req_fields, crit_fields, result, key)
-
-        # After batch processing
-        shared_cache.shutdown()
-    """
-
-    def __init__(self, max_size: int = 1000, ttl_seconds: int = 3600, logger: logging.Logger = None):
-        """
-        Initialize shared validation cache.
-
-        Args:
-            max_size: Maximum number of cached entries (default: 1000)
-            ttl_seconds: Time-to-live for cache entries in seconds (default: 3600)
-            logger: Logger instance for cache statistics (default: module logger)
-        """
-        self.max_size = max_size
-        self.ttl_seconds = ttl_seconds
-        self.logger = logger or logging.getLogger(__name__)
-
-        # Create multiprocessing Manager for shared state
-        self._manager = multiprocessing.Manager()
-
-        # Shared cache storage: key -> (issues_list, timestamp)
-        self._cache = self._manager.dict()
-
-        # Shared LRU tracking: key -> last_access_time
-        self._access_times = self._manager.dict()
-
-        # Shared statistics
-        self._stats = self._manager.dict({
-            'hits': 0,
-            'misses': 0,
-            'evictions': 0
-        })
-
-        # Manager-level lock for cache operations
-        self._lock = self._manager.Lock()
-
-    def _generate_cache_key(self, df: pd.DataFrame, item_type: str,
-                            required_fields: List[str], critical_fields: List[str]) -> str:
-        """
-        Generate cache key from DataFrame content and configuration.
-
-        Same algorithm as ValidationCache for compatibility.
-        """
-        try:
-            # Hash DataFrame structure and content
-            df_hash = pd.util.hash_pandas_object(df, index=False).sum()
-
-            # Hash configuration (required_fields + critical_fields)
-            config_str = f"{sorted(required_fields)}:{sorted(critical_fields)}"
-            config_hash = hashlib.md5(config_str.encode()).hexdigest()[:8]
-
-            # Combine into cache key
-            return f"{item_type}:{df_hash}:{config_hash}"
-
-        except Exception as e:
-            self.logger.warning(f"Error generating cache key: {e}. Cache disabled for this call.")
-            # Return unique key to force cache miss
-            return f"error:{time.time()}"
-
-    def get(self, df: pd.DataFrame, item_type: str,
-            required_fields: List[str], critical_fields: List[str]) -> Tuple[Optional[List[Dict]], str]:
-        """
-        Retrieve cached validation results if available.
-
-        Same interface as ValidationCache for compatibility.
-
-        Args:
-            df: DataFrame to look up in cache
-            item_type: 'Metrics' or 'Dimensions'
-            required_fields: List of required field names
-            critical_fields: List of critical field names
-
-        Returns:
-            Tuple of (issues list or None, cache_key).
-            The cache_key can be passed to put() to avoid recomputing the hash.
-        """
-        cache_key = self._generate_cache_key(df, item_type, required_fields, critical_fields)
-
-        with self._lock:
-            if cache_key not in self._cache:
-                self._stats['misses'] = self._stats.get('misses', 0) + 1
-                return None, cache_key
-
-            cached_data = self._cache[cache_key]
-            cached_issues, timestamp = cached_data
-
-            # Check TTL expiration
-            age = time.time() - timestamp
-            if age > self.ttl_seconds:
-                del self._cache[cache_key]
-                if cache_key in self._access_times:
-                    del self._access_times[cache_key]
-                self._stats['misses'] = self._stats.get('misses', 0) + 1
-                return None, cache_key
-
-            # Cache hit - update access time and stats
-            self._access_times[cache_key] = time.time()
-            self._stats['hits'] = self._stats.get('hits', 0) + 1
-
-            # Return copy to prevent mutation
-            return [issue.copy() for issue in cached_issues], cache_key
-
-    def put(self, df: pd.DataFrame, item_type: str,
-            required_fields: List[str], critical_fields: List[str],
-            issues: List[Dict], cache_key: str = None) -> None:
-        """
-        Store validation results in cache.
-
-        Same interface as ValidationCache for compatibility.
-
-        Args:
-            df: DataFrame being cached (used if cache_key not provided)
-            item_type: 'Metrics' or 'Dimensions'
-            required_fields: List of required field names
-            critical_fields: List of critical field names
-            issues: List of validation issues to cache
-            cache_key: Optional pre-computed cache key from get()
-        """
-        if cache_key is None:
-            cache_key = self._generate_cache_key(df, item_type, required_fields, critical_fields)
-
-        with self._lock:
-            # Evict oldest entry if cache is full
-            if len(self._cache) >= self.max_size and cache_key not in self._cache:
-                self._evict_lru()
-
-            # Store issues with timestamp (deep copy for safety)
-            self._cache[cache_key] = ([issue.copy() for issue in issues], time.time())
-            self._access_times[cache_key] = time.time()
-
-    def _evict_lru(self) -> None:
-        """Evict least recently used cache entry (must be called within lock)."""
-        if not self._access_times:
-            return
-
-        # Find least recently used key
-        access_times_dict = dict(self._access_times)
-        if not access_times_dict:
-            return
-
-        lru_key = min(access_times_dict.items(), key=lambda x: x[1])[0]
-
-        # Remove from cache
-        if lru_key in self._cache:
-            del self._cache[lru_key]
-        if lru_key in self._access_times:
-            del self._access_times[lru_key]
-
-        self._stats['evictions'] = self._stats.get('evictions', 0) + 1
-
-    def get_statistics(self) -> Dict[str, Any]:
-        """
-        Get cache performance statistics.
-
-        Returns:
-            Dict with hits, misses, hit_rate, size, evictions
-        """
-        with self._lock:
-            stats = dict(self._stats)
-            hits = stats.get('hits', 0)
-            misses = stats.get('misses', 0)
-            total_requests = hits + misses
-            hit_rate = (hits / total_requests * 100) if total_requests > 0 else 0
-
-            return {
-                'hits': hits,
-                'misses': misses,
-                'hit_rate': hit_rate,
-                'size': len(self._cache),
-                'max_size': self.max_size,
-                'evictions': stats.get('evictions', 0),
-                'total_requests': total_requests
-            }
-
-    def clear(self) -> None:
-        """Clear all cache entries."""
-        with self._lock:
-            self._cache.clear()
-            self._access_times.clear()
-
-    def shutdown(self) -> None:
-        """
-        Shutdown the Manager and cleanup resources.
-
-        IMPORTANT: Call this after batch processing is complete to avoid
-        resource leaks. The cache cannot be used after shutdown.
-        """
-        try:
-            self._manager.shutdown()
-        except Exception:
-            pass  # Manager may already be shutdown
-
+from cja_auto_sdr.api.cache import ValidationCache, SharedValidationCache  # noqa: F811
 
 # ==================== PROFILE MANAGEMENT ====================
 
@@ -1756,1012 +917,21 @@ def test_profile(profile_name: str) -> bool:
         return False
 
 
-# ==================== CONFIG VALIDATION HELPERS ====================
+# ==================== CONFIG VALIDATION (moved to core/config_validation.py) ====================
+from cja_auto_sdr.core.config_validation import ConfigValidator, validate_credentials, validate_config_file  # noqa: F811,E402
 
-class ConfigValidator:
-    """Provides detailed validation and suggestions for configuration fields."""
+# ==================== CREDENTIAL LOADING (moved to core/credentials.py) ====================
+from cja_auto_sdr.core.credentials import (  # noqa: F811,E402
+    CredentialLoader, JsonFileCredentialLoader, DotenvCredentialLoader,
+    EnvironmentCredentialLoader, CredentialResolver,
+    normalize_credential_value, filter_credentials,
+    load_credentials_from_env, validate_env_credentials,
+)
 
-    @staticmethod
-    def validate_org_id(org_id: str) -> Tuple[bool, Optional[str]]:
-        """
-        Validate that ORG_ID has the correct format.
+# ==================== CJA CLIENT (moved to api/client.py) ====================
+from cja_auto_sdr.api.client import configure_cjapy, initialize_cja, _config_from_env  # noqa: F811,E402
 
-        Args:
-            org_id: The organization ID to validate
 
-        Returns:
-            Tuple of (is_valid, error_message). error_message is None if valid.
-        """
-        if not org_id or not org_id.strip():
-            return False, "ORG_ID cannot be empty"
-
-        org_id = org_id.strip()
-
-        # Check for @AdobeOrg suffix
-        if not org_id.endswith('@AdobeOrg'):
-            # Try to be helpful by detecting common mistakes
-            if '@' in org_id:
-                return False, (
-                    f"ORG_ID '{org_id}' has incorrect suffix. "
-                    f"It must end with '@AdobeOrg', not '{org_id.split('@')[1]}'"
-                )
-            return False, (
-                f"ORG_ID '{org_id}' is missing '@AdobeOrg' suffix. "
-                f"Correct format: '{org_id}@AdobeOrg'"
-            )
-
-        # Check that there's something before the @AdobeOrg
-        org_prefix = org_id[:-9]  # Remove '@AdobeOrg'
-        if not org_prefix:
-            return False, "ORG_ID cannot be just '@AdobeOrg' - needs organization prefix"
-
-        return True, None
-
-    @staticmethod
-    def validate_scopes(scopes: str) -> Tuple[bool, Optional[str], List[str]]:
-        """
-        Validate OAuth scopes are provided.
-
-        Scopes vary based on Adobe Developer Console project configuration.
-        See: https://developer.adobe.com/developer-console/docs/guides/authentication/
-
-        Args:
-            scopes: Comma or space-separated scopes string
-
-        Returns:
-            Tuple of (is_valid, error_message, missing_scopes).
-            error_message is None if valid. missing_scopes is always empty (no longer validated).
-        """
-        if not scopes or not scopes.strip():
-            return False, "SCOPES cannot be empty - copy from Adobe Developer Console", []
-
-        return True, None, []
-
-    @staticmethod
-    def validate_client_id(client_id: str) -> Tuple[bool, Optional[str]]:
-        """
-        Validate client ID format.
-
-        Args:
-            client_id: The client ID to validate
-
-        Returns:
-            Tuple of (is_valid, error_message). error_message is None if valid.
-        """
-        if not client_id or not client_id.strip():
-            return False, "CLIENT_ID cannot be empty"
-
-        client_id = client_id.strip()
-
-        # Adobe client IDs are typically 32 hex characters
-        if len(client_id) < 16:
-            return False, (
-                f"CLIENT_ID '{client_id[:8]}...' appears too short. "
-                f"Adobe OAuth client IDs are typically 32 characters."
-            )
-
-        return True, None
-
-    @staticmethod
-    def validate_secret(secret: str) -> Tuple[bool, Optional[str]]:
-        """
-        Validate client secret format.
-
-        Args:
-            secret: The client secret to validate
-
-        Returns:
-            Tuple of (is_valid, error_message). error_message is None if valid.
-        """
-        if not secret or not secret.strip():
-            return False, "SECRET cannot be empty"
-
-        secret = secret.strip()
-
-        # Adobe secrets are typically longer
-        if len(secret) < 16:
-            return False, "SECRET appears too short. Adobe OAuth secrets are typically longer."
-
-        return True, None
-
-    @classmethod
-    def validate_all(cls, credentials: Dict[str, str], logger: logging.Logger) -> List[str]:
-        """
-        Run all validations and return list of issues.
-
-        Args:
-            credentials: Dictionary of credentials
-            logger: Logger instance
-
-        Returns:
-            List of validation issues (empty if all valid)
-        """
-        issues = []
-
-        # Validate ORG_ID
-        if 'org_id' in credentials:
-            valid, error = cls.validate_org_id(credentials['org_id'])
-            if not valid:
-                issues.append(error)
-                logger.warning(f"Configuration issue: {error}")
-
-        # Validate CLIENT_ID
-        if 'client_id' in credentials:
-            valid, error = cls.validate_client_id(credentials['client_id'])
-            if not valid:
-                issues.append(error)
-                logger.warning(f"Configuration issue: {error}")
-
-        # Validate SECRET
-        if 'secret' in credentials:
-            valid, error = cls.validate_secret(credentials['secret'])
-            if not valid:
-                issues.append(error)
-                logger.warning(f"Configuration issue: {error}")
-
-        # Validate SCOPES (warning only - not strictly required)
-        if 'scopes' in credentials:
-            valid, error, missing = cls.validate_scopes(credentials['scopes'])
-            if not valid:
-                logger.warning(f"Configuration warning: {error}")
-                # Don't add to issues - scopes are a warning, not an error
-
-        return issues
-
-
-def validate_credentials(
-    credentials: Dict[str, str],
-    logger: logging.Logger,
-    strict: bool = False,
-    source: str = "unknown"
-) -> Tuple[bool, List[str]]:
-    """Unified validation for credentials from any source.
-
-    Provides consistent validation across profiles, environment variables,
-    and config files using CONFIG_SCHEMA and ConfigValidator.
-
-    Args:
-        credentials: Dictionary of credentials to validate
-        logger: Logger instance
-        strict: If True, fail on any issue. If False, return issues list but may pass.
-        source: Name of credential source for logging (e.g., "profile", "env", "config_file")
-
-    Returns:
-        Tuple of (is_valid, issues). is_valid is True if credentials are usable.
-    """
-    issues = []
-
-    # Check required fields are present and non-empty
-    for field in CREDENTIAL_FIELDS['required']:
-        if field not in credentials:
-            issues.append(f"Missing required field: '{field}'")
-        elif not credentials[field] or not str(credentials[field]).strip():
-            issues.append(f"Empty value for required field: '{field}'")
-
-    # Run detailed validation with ConfigValidator
-    validation_issues = ConfigValidator.validate_all(credentials, logger)
-    issues.extend(validation_issues)
-
-    # Check for scopes (warning, not error)
-    if 'scopes' not in credentials or not credentials.get('scopes', '').strip():
-        logger.warning(
-            f"Credentials from {source} missing OAuth scopes - "
-            "recommend setting scopes (copy from Adobe Developer Console)"
-        )
-
-    # Filter credentials to known fields only
-    unknown_fields = set(credentials.keys()) - CREDENTIAL_FIELDS['all']
-    if unknown_fields:
-        logger.debug(f"Ignoring unknown fields from {source}: {', '.join(unknown_fields)}")
-
-    is_valid = len(issues) == 0 if strict else all(
-        "Missing required field" not in issue and "Empty value" not in issue
-        for issue in issues
-    )
-
-    if issues:
-        for issue in issues:
-            logger.warning(f"Credential validation ({source}): {issue}")
-
-    return is_valid, issues
-
-
-def normalize_credential_value(value: Any) -> str:
-    """Normalize a credential value consistently across all sources.
-
-    Handles stripping whitespace and quotes from values.
-
-    Args:
-        value: The value to normalize (can be any type)
-
-    Returns:
-        Normalized string value
-    """
-    if value is None:
-        return ""
-    s = str(value).strip()
-    # Remove surrounding quotes (common in .env files)
-    if len(s) >= 2 and ((s.startswith('"') and s.endswith('"')) or
-                         (s.startswith("'") and s.endswith("'"))):
-        s = s[1:-1]
-    return s
-
-
-def filter_credentials(credentials: Dict[str, Any]) -> Dict[str, str]:
-    """Filter and normalize credentials to known fields only.
-
-    Args:
-        credentials: Raw credentials dictionary
-
-    Returns:
-        Filtered dictionary with only known credential fields, normalized values
-    """
-    return {
-        k: normalize_credential_value(v)
-        for k, v in credentials.items()
-        if k in CREDENTIAL_FIELDS['all'] and v
-    }
-
-
-# ==================== CREDENTIAL LOADERS ====================
-
-class CredentialLoader(ABC):
-    """Abstract base class for loading credentials from different sources.
-
-    Provides consistent interface and error handling for all credential sources.
-    """
-
-    @property
-    @abstractmethod
-    def source_name(self) -> str:
-        """Return the name of this credential source."""
-        pass
-
-    def load(self, logger: logging.Logger) -> Optional[Dict[str, str]]:
-        """Load credentials, handling errors gracefully.
-
-        Args:
-            logger: Logger instance
-
-        Returns:
-            Dictionary of credentials if successful, None otherwise
-        """
-        try:
-            creds = self._load_impl(logger)
-            if creds:
-                return filter_credentials(creds)
-            return None
-        except (IOError, json.JSONDecodeError) as e:
-            logger.debug(f"Failed to load credentials from {self.source_name}: {e}")
-            return None
-
-    @abstractmethod
-    def _load_impl(self, logger: logging.Logger) -> Optional[Dict[str, Any]]:
-        """Implementation-specific loading logic.
-
-        Args:
-            logger: Logger instance
-
-        Returns:
-            Raw credentials dictionary, or None if not available
-        """
-        pass
-
-
-class JsonFileCredentialLoader(CredentialLoader):
-    """Load credentials from a JSON file."""
-
-    def __init__(self, file_path: Path):
-        self.file_path = file_path
-
-    @property
-    def source_name(self) -> str:
-        return f"json:{self.file_path.name}"
-
-    def _load_impl(self, logger: logging.Logger) -> Optional[Dict[str, Any]]:
-        if not self.file_path.exists():
-            return None
-
-        with open(self.file_path, 'r') as f:
-            config = json.load(f)
-
-        if isinstance(config, dict):
-            return config
-        return None
-
-
-class DotenvCredentialLoader(CredentialLoader):
-    """Load credentials from a .env file."""
-
-    def __init__(self, file_path: Path):
-        self.file_path = file_path
-
-    @property
-    def source_name(self) -> str:
-        return f"dotenv:{self.file_path.name}"
-
-    def _load_impl(self, logger: logging.Logger) -> Optional[Dict[str, Any]]:
-        if not self.file_path.exists():
-            return None
-
-        credentials = {}
-        with open(self.file_path, 'r') as f:
-            for line in f:
-                line = line.strip()
-                if not line or line.startswith('#'):
-                    continue
-                if '=' in line:
-                    key, _, value = line.partition('=')
-                    key = key.strip().lower()
-                    value = value.strip()
-                    # Remove quotes
-                    if value and len(value) >= 2:
-                        if (value.startswith('"') and value.endswith('"')) or \
-                           (value.startswith("'") and value.endswith("'")):
-                            value = value[1:-1]
-                    if key and value:
-                        credentials[key] = value
-
-        return credentials if credentials else None
-
-
-class EnvironmentCredentialLoader(CredentialLoader):
-    """Load credentials from environment variables."""
-
-    @property
-    def source_name(self) -> str:
-        return "environment"
-
-    def _load_impl(self, logger: logging.Logger) -> Optional[Dict[str, Any]]:
-        credentials = {}
-        for config_key, env_var in ENV_VAR_MAPPING.items():
-            value = os.environ.get(env_var)
-            if value and value.strip():
-                credentials[config_key] = value.strip()
-
-        return credentials if credentials else None
-
-
-# ==================== CREDENTIAL RESOLVER ====================
-
-class CredentialResolver:
-    """Resolves credentials using priority-based strategy.
-
-    Priority order:
-    1. Profile credentials (if --profile or CJA_PROFILE specified)
-    2. Environment variables
-    3. Configuration file
-
-    This class consolidates the credential loading logic that was previously
-    scattered throughout initialize_cja().
-    """
-
-    def __init__(self, logger: logging.Logger):
-        self.logger = logger
-
-    def resolve(
-        self,
-        profile: Optional[str] = None,
-        config_file: Union[str, Path] = "config.json"
-    ) -> Tuple[Dict[str, str], str]:
-        """Resolve credentials following priority order.
-
-        Args:
-            profile: Profile name to load (optional)
-            config_file: Path to config file (fallback)
-
-        Returns:
-            Tuple of (credentials, source_name)
-
-        Raises:
-            CredentialSourceError: If no valid credentials found
-        """
-        # 1. Try profile credentials (highest priority)
-        if profile:
-            creds, source = self._try_profile(profile)
-            if creds:
-                return creds, source
-
-        # 2. Try environment variables
-        creds, source = self._try_environment()
-        if creds:
-            # Warn if config file also exists
-            config_path = Path(config_file)
-            if config_path.exists():
-                self._warn_multiple_sources(config_path)
-            return creds, source
-
-        # 3. Try config file (lowest priority)
-        creds, source = self._try_config_file(config_file)
-        if creds:
-            return creds, source
-
-        # No valid credentials found
-        raise CredentialSourceError(
-            "No valid credentials found",
-            source="all",
-            reason="Checked profile, environment variables, and config file",
-            details="Set credentials via --profile, environment variables, or config.json"
-        )
-
-    def _try_profile(self, profile_name: str) -> Tuple[Optional[Dict[str, str]], str]:
-        """Try to load credentials from a profile.
-
-        Args:
-            profile_name: Name of the profile
-
-        Returns:
-            Tuple of (credentials or None, source_name)
-
-        Raises:
-            CredentialSourceError: If profile loading fails critically
-        """
-        self.logger.info(f"Loading credentials from profile '{profile_name}'...")
-        try:
-            creds = load_profile_credentials(profile_name, self.logger)
-            if creds:
-                is_valid, issues = validate_credentials(
-                    creds, self.logger, strict=False, source=f"profile:{profile_name}"
-                )
-                if is_valid:
-                    self.logger.info(f"Using credentials from profile '{profile_name}'")
-                    return creds, f"profile:{profile_name}"
-                else:
-                    self.logger.warning(
-                        f"Profile '{profile_name}' credentials have issues: {issues}"
-                    )
-        except ProfileNotFoundError as e:
-            raise CredentialSourceError(
-                str(e),
-                source=f"profile:{profile_name}",
-                reason="Profile directory not found",
-                details=e.details
-            )
-        except ProfileConfigError as e:
-            raise CredentialSourceError(
-                str(e),
-                source=f"profile:{profile_name}",
-                reason="Invalid profile configuration",
-                details=e.details
-            )
-
-        return None, ""
-
-    def _try_environment(self) -> Tuple[Optional[Dict[str, str]], str]:
-        """Try to load credentials from environment variables.
-
-        Returns:
-            Tuple of (credentials or None, source_name)
-        """
-        loader = EnvironmentCredentialLoader()
-        creds = loader.load(self.logger)
-
-        if creds:
-            is_valid, _ = validate_credentials(
-                creds, self.logger, strict=False, source="environment"
-            )
-            if is_valid:
-                self.logger.info("Using credentials from environment variables")
-                return creds, "environment"
-            else:
-                self.logger.debug("Environment credentials incomplete, trying next source")
-
-        return None, ""
-
-    def _try_config_file(
-        self,
-        config_file: Union[str, Path]
-    ) -> Tuple[Optional[Dict[str, str]], str]:
-        """Try to load credentials from config file.
-
-        Args:
-            config_file: Path to config file
-
-        Returns:
-            Tuple of (credentials or None, source_name)
-
-        Raises:
-            CredentialSourceError: If config file has critical errors
-        """
-        config_path = Path(config_file)
-
-        if not config_path.exists():
-            self.logger.debug(f"Config file not found: {config_path}")
-            return None, ""
-
-        loader = JsonFileCredentialLoader(config_path)
-        creds = loader.load(self.logger)
-
-        if creds:
-            is_valid, issues = validate_credentials(
-                creds, self.logger, strict=False, source=f"config:{config_path.name}"
-            )
-            if is_valid:
-                self.logger.info(f"Using credentials from {config_file}")
-                return creds, f"config:{config_path.name}"
-            else:
-                # Config file exists but has issues
-                raise CredentialSourceError(
-                    f"Config file '{config_file}' has validation errors",
-                    source=f"config:{config_path.name}",
-                    reason="; ".join(issues[:3]),  # Show first 3 issues
-                    details="Fix the issues or use environment variables instead"
-                )
-
-        return None, ""
-
-    def _warn_multiple_sources(self, config_path: Path) -> None:
-        """Warn user when multiple credential sources exist."""
-        self.logger.warning("=" * BANNER_WIDTH)
-        self.logger.warning("NOTICE: Both environment variables AND config file detected")
-        self.logger.warning(f"  Environment variables: ORG_ID, CLIENT_ID, SECRET, etc.")
-        self.logger.warning(f"  Config file: {config_path}")
-        self.logger.warning("  Using: ENVIRONMENT VARIABLES (takes precedence)")
-        self.logger.warning("")
-        self.logger.warning("To avoid confusion:")
-        self.logger.warning("  - Remove config.json if using environment variables")
-        self.logger.warning("  - Or unset env vars: unset ORG_ID CLIENT_ID SECRET SCOPES")
-        self.logger.warning("=" * BANNER_WIDTH)
-
-
-def load_credentials_from_env() -> Optional[Dict[str, str]]:
-    """
-    Load Adobe API credentials from environment variables.
-
-    Environment variables:
-        ORG_ID: Adobe Organization ID
-        CLIENT_ID: OAuth Client ID
-        SECRET: Client Secret
-        SCOPES: OAuth scopes
-        SANDBOX: Sandbox name (optional)
-
-    Returns:
-        Dictionary with credentials if any env vars are set, None otherwise
-    """
-    credentials = {}
-    for config_key, env_var in ENV_VAR_MAPPING.items():
-        value = os.environ.get(env_var)
-        if value and value.strip():
-            credentials[config_key] = value.strip()
-
-    # Return None if no CJA environment variables are set
-    if not credentials:
-        return None
-
-    return credentials
-
-
-def validate_env_credentials(credentials: Dict[str, str], logger: logging.Logger) -> bool:
-    """
-    Validate that environment credentials have minimum required fields for OAuth.
-
-    Uses CREDENTIAL_FIELDS (single source of truth) and validate_credentials for
-    consistent validation across all credential sources.
-
-    Args:
-        credentials: Dictionary of credentials from environment
-        logger: Logger instance
-
-    Returns:
-        True if credentials have minimum required fields
-    """
-    # Use CREDENTIAL_FIELDS for required fields (single source of truth)
-    for field in CREDENTIAL_FIELDS['required']:
-        if field not in credentials or not credentials[field].strip():
-            logger.debug(f"Missing required environment variable: {ENV_VAR_MAPPING.get(field, field)}")
-            return False
-
-    # Use unified validation function
-    is_valid, issues = validate_credentials(credentials, logger, strict=False, source="environment")
-
-    return is_valid
-
-
-def _config_from_env(credentials: Dict[str, str], logger: logging.Logger):
-    """
-    Configure cjapy using environment credentials.
-
-    Creates a temporary JSON config file that is cleaned up on exit.
-
-    Args:
-        credentials: Dictionary of credentials from environment
-        logger: Logger instance
-    """
-    # cjapy.importConfigFile expects a JSON file, so we create a temporary one
-    # This is cleaned up on exit. Use restrictive permissions (0o600) since it contains credentials.
-    temp_config = tempfile.NamedTemporaryFile(
-        mode='w',
-        suffix='.json',
-        delete=False,
-        prefix='cja_env_config_'
-    )
-    json.dump(credentials, temp_config)
-    temp_config.close()
-    os.chmod(temp_config.name, 0o600)
-
-    logger.debug(f"Created temporary config file: {temp_config.name}")
-
-    # Register cleanup
-    def cleanup_temp_config():
-        try:
-            os.unlink(temp_config.name)
-        except OSError:
-            pass
-
-    atexit.register(cleanup_temp_config)
-
-    # Import the temporary config
-    cjapy.importConfigFile(temp_config.name)
-
-
-def configure_cjapy(
-    profile: Optional[str] = None,
-    config_file: str = "config.json",
-    logger: Optional[logging.Logger] = None
-) -> Tuple[bool, str, Optional[Dict[str, str]]]:
-    """
-    Configure cjapy with credentials using priority: profile > env > config file.
-
-    This is a lightweight configuration function that sets up cjapy credentials
-    without creating a CJA instance. Use this for commands that need credential
-    configuration but manage their own CJA lifecycle.
-
-    Uses CredentialResolver internally to resolve credentials following the
-    standard priority order.
-
-    Args:
-        profile: Optional profile name to use for credentials
-        config_file: Path to fallback config file
-        logger: Optional logger instance
-
-    Returns:
-        Tuple of (success, source_description, credentials_dict)
-        - success: True if cjapy was configured successfully
-        - source_description: Description of credential source used
-        - credentials_dict: The credentials that were used (for display purposes)
-    """
-    if logger is None:
-        logger = logging.getLogger(__name__)
-        logger.setLevel(logging.WARNING)
-
-    try:
-        # Use CredentialResolver for unified credential resolution
-        resolver = CredentialResolver(logger)
-        credentials, source = resolver.resolve(profile=profile, config_file=config_file)
-
-        # Configure cjapy with the resolved credentials
-        # Source format from resolver: "profile:name", "environment", "config:filename"
-        if source.startswith("profile:") or source == "environment":
-            # Profile or environment credentials need temp file for cjapy
-            _config_from_env(credentials, logger)
-        else:
-            # Config file can be imported directly
-            cjapy.importConfigFile(config_file)
-
-        # Format source for display (convert resolver format to user-friendly format)
-        if source.startswith("profile:"):
-            display_source = f"Profile: {source.split(':', 1)[1]}"
-        elif source == "environment":
-            display_source = "Environment variables"
-        elif source.startswith("config:"):
-            display_source = f"Config file: {Path(config_file).resolve()}"
-        else:
-            display_source = source
-
-        return True, display_source, credentials
-
-    except CredentialSourceError as e:
-        logger.error(f"Credential error: {e}")
-        return False, str(e), None
-    except (ProfileNotFoundError, ProfileConfigError) as e:
-        logger.error(f"Profile error: {e}")
-        return False, f"Profile error: {e}", None
-
-
-# ==================== CJA INITIALIZATION ====================
-
-def validate_config_file(
-    config_file: Union[str, Path],
-    logger: logging.Logger
-) -> bool:
-    """
-    Validate configuration file exists and has required structure.
-
-    Performs comprehensive validation:
-    1. File existence and readability
-    2. JSON syntax validation
-    3. Required fields presence
-    4. Field type validation
-    5. Empty value detection
-    6. Private key file validation (if path provided)
-
-    Args:
-        config_file: Path to the configuration JSON file
-        logger: Logger instance for output
-
-    Returns:
-        True if validation passes, False otherwise
-
-    Raises:
-        ConfigurationError: If validation fails (when exceptions are preferred)
-    """
-    validation_errors = []
-    validation_warnings = []
-
-    try:
-        logger.info(f"Validating configuration file: {config_file}")
-
-        config_path = Path(config_file)
-
-        # Check if file exists
-        if not config_path.exists():
-            error_msg = ErrorMessageHelper.get_config_error_message(
-                "file_not_found",
-                details=f"Looking for: {config_path.absolute()}"
-            )
-            logger.error("\n" + error_msg)
-            return False
-
-        # Check if file is readable
-        if not config_path.is_file():
-            logger.error(f"'{config_file}' is not a valid file")
-            return False
-
-        # Validate JSON structure
-        try:
-            with open(config_path, 'r') as f:
-                config_data = json.load(f)
-        except json.JSONDecodeError as e:
-            error_msg = ErrorMessageHelper.get_config_error_message(
-                "invalid_json",
-                details=f"Line {e.lineno}, Column {e.colno}: {e.msg}"
-            )
-            logger.error("\n" + error_msg)
-            return False
-
-        # Validate it's a dictionary
-        if not isinstance(config_data, dict):
-            logger.error("Configuration file must contain a JSON object (dictionary)")
-            return False
-
-        # Check for base required fields (required for all auth methods)
-        for field_name, field_info in CONFIG_SCHEMA['base_required_fields'].items():
-            if field_name not in config_data:
-                validation_errors.append(f"Missing required field: '{field_name}' ({field_info['description']})")
-            elif not isinstance(config_data[field_name], field_info['type']):
-                validation_errors.append(
-                    f"Invalid type for '{field_name}': expected {field_info['type'].__name__}, "
-                    f"got {type(config_data[field_name]).__name__}"
-                )
-            elif not config_data[field_name] or (isinstance(config_data[field_name], str) and not config_data[field_name].strip()):
-                validation_errors.append(f"Empty value for required field: '{field_name}'")
-
-        # OAuth Server-to-Server auth - warn if scopes not provided
-        if 'scopes' not in config_data or not config_data.get('scopes', '').strip():
-            validation_warnings.append(
-                "OAuth Server-to-Server auth: 'scopes' field not set. "
-                "Copy scopes from your Adobe Developer Console project."
-            )
-
-        # Validate optional fields if present
-        for field_name, field_info in CONFIG_SCHEMA['optional_fields'].items():
-            if field_name in config_data:
-                if not isinstance(config_data[field_name], field_info['type']):
-                    validation_warnings.append(
-                        f"Invalid type for optional field '{field_name}': expected {field_info['type'].__name__}"
-                    )
-
-        # Check for deprecated JWT authentication fields
-        deprecated_found = []
-        for field, description in JWT_DEPRECATED_FIELDS.items():
-            if field in config_data:
-                deprecated_found.append(f"'{field}' ({description})")
-        if deprecated_found:
-            validation_warnings.append(
-                f"DEPRECATED: JWT authentication was removed in v3.0.8. "
-                f"Found JWT fields: {', '.join(deprecated_found)}. "
-                f"Please migrate to OAuth Server-to-Server authentication. "
-                f"See docs/QUICKSTART_GUIDE.md for setup instructions."
-            )
-
-        # Check for unknown fields (potential typos)
-        known_fields = (set(CONFIG_SCHEMA['base_required_fields'].keys()) |
-                        set(CONFIG_SCHEMA['optional_fields'].keys()) |
-                        set(JWT_DEPRECATED_FIELDS.keys()))  # Include deprecated fields as "known"
-        unknown_fields = set(config_data.keys()) - known_fields
-        if unknown_fields:
-            validation_warnings.append(f"Unknown fields in config (possible typos): {', '.join(unknown_fields)}")
-
-        # Report validation results
-        if validation_errors:
-            logger.error("Configuration validation FAILED:")
-            for error in validation_errors:
-                logger.error(f"  - {error}")
-            logger.error("")
-
-            # Provide enhanced error message if missing credentials
-            if any("Missing required field" in err for err in validation_errors):
-                error_msg = ErrorMessageHelper.get_config_error_message(
-                    "missing_credentials",
-                    details="One or more required fields are missing from your config file"
-                )
-                logger.error(error_msg)
-            elif any("Empty value" in err for err in validation_errors):
-                error_msg = ErrorMessageHelper.get_config_error_message(
-                    "invalid_format",
-                    details="One or more fields have empty or invalid values"
-                )
-                logger.error(error_msg)
-            return False
-
-        if validation_warnings:
-            logger.warning("Configuration validation warnings:")
-            for warning in validation_warnings:
-                logger.warning(f"  - {warning}")
-
-        logger.info("Configuration file validated successfully")
-        return True
-
-    except PermissionError as e:
-        logger.error(f"Permission denied reading config file: {e}")
-        logger.error("Check file permissions for the configuration file")
-        return False
-    except Exception as e:
-        logger.error(f"Unexpected error validating config file ({type(e).__name__}): {str(e)}")
-        return False
-
-def initialize_cja(
-    config_file: Union[str, Path] = "config.json",
-    logger: Optional[logging.Logger] = None,
-    profile: Optional[str] = None
-) -> Optional[cjapy.CJA]:
-    """Initialize CJA connection with comprehensive error handling.
-
-    Uses CredentialResolver for unified credential loading with priority:
-        1. Profile credentials (if --profile or CJA_PROFILE specified)
-        2. Environment variables (ORG_ID, CLIENT_ID, SECRET, etc.)
-        3. Configuration file (config.json)
-
-    Args:
-        config_file: Path to CJA configuration file
-        logger: Logger instance (uses module logger if None)
-        profile: Profile name to load credentials from (optional)
-
-    Returns:
-        Initialized CJA instance, or None if initialization fails
-
-    Raises:
-        ConfigurationError: If credentials are invalid or missing
-        APIError: If API connection fails
-    """
-    logger = logger or logging.getLogger(__name__)
-    try:
-        logger.info("=" * BANNER_WIDTH)
-        logger.info("INITIALIZING CJA CONNECTION")
-        logger.info("=" * BANNER_WIDTH)
-
-        # Log dotenv status for debugging
-        if _DOTENV_AVAILABLE:
-            if _DOTENV_LOADED:
-                logger.debug(".env file found and loaded")
-            else:
-                logger.debug(".env file not found (python-dotenv available but no .env file)")
-        else:
-            logger.debug("python-dotenv not installed (.env files will not be auto-loaded)")
-
-        # Resolve active profile (--profile > CJA_PROFILE > None)
-        active_profile = resolve_active_profile(profile)
-
-        # Use CredentialResolver for unified credential loading
-        resolver = CredentialResolver(logger)
-        try:
-            credentials, source = resolver.resolve(
-                profile=active_profile,
-                config_file=config_file
-            )
-            logger.info(f"Credentials loaded from: {source}")
-        except CredentialSourceError as e:
-            logger.critical("=" * BANNER_WIDTH)
-            logger.critical("CREDENTIAL LOADING FAILED")
-            logger.critical("=" * BANNER_WIDTH)
-            logger.critical(str(e))
-            if e.reason:
-                logger.critical(f"Reason: {e.reason}")
-            if e.details:
-                logger.critical(e.details)
-            logger.critical("")
-            logger.critical("Options to provide credentials:")
-            logger.critical("  1. Profile:   cja_auto_sdr --profile <name> ...")
-            logger.critical("  2. Env vars:  export ORG_ID=... CLIENT_ID=... SECRET=... SCOPES=...")
-            logger.critical("  3. Config:    Create config.json with credentials")
-            return None
-
-        # Configure cjapy with resolved credentials
-        if source.startswith("config:"):
-            # Config file - use cjapy's native loading
-            logger.info(f"Loading CJA configuration from {config_file}...")
-            cjapy.importConfigFile(config_file)
-        else:
-            # Profile or environment - use temp config file
-            logger.info(f"Loading CJA configuration from {source}...")
-            _config_from_env(credentials, logger)
-
-        logger.info("Configuration loaded successfully")
-
-        # Attempt to create CJA instance
-        logger.info("Creating CJA instance...")
-        cja = cjapy.CJA()
-        logger.info("CJA instance created successfully")
-
-        # Test connection with a simple API call (with retry)
-        logger.info("Testing API connection...")
-        try:
-            # Attempt to list data views to verify connection with retry
-            test_call = make_api_call_with_retry(
-                cja.getDataViews,
-                logger=logger,
-                operation_name="getDataViews (connection test)"
-            )
-            if test_call is not None:
-                logger.info(f"✓ API connection successful! Found {len(test_call) if hasattr(test_call, '__len__') else 'multiple'} data view(s)")
-            else:
-                logger.warning("API connection test returned None - connection may be unstable")
-        except Exception as test_error:
-            logger.warning(f"Could not verify connection with test call: {str(test_error)}")
-            logger.warning("Proceeding anyway - errors may occur during data fetching")
-
-        logger.info("CJA initialization complete")
-        return cja
-
-    except FileNotFoundError as e:
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical("CONFIGURATION FILE ERROR")
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical(f"Config file not found: {config_file}")
-        logger.critical(f"Current working directory: {Path.cwd()}")
-        logger.critical("Please ensure the configuration file exists in the correct location")
-        return None
-
-    except ImportError as e:
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical("DEPENDENCY ERROR")
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical(f"Failed to import cjapy module: {str(e)}")
-        logger.critical("Please ensure cjapy is installed: pip install cjapy")
-        return None
-
-    except AttributeError as e:
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical("CJA CONFIGURATION ERROR")
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical(f"Configuration error: {str(e)}")
-        logger.critical("This usually indicates an issue with the authentication credentials")
-        logger.critical("Please verify all fields in your configuration file are correct")
-        return None
-
-    except PermissionError as e:
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical("PERMISSION ERROR")
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical(f"Cannot read configuration file: {str(e)}")
-        logger.critical("Please check file permissions")
-        return None
-
-    except Exception as e:
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical("CJA INITIALIZATION FAILED")
-        logger.critical("=" * BANNER_WIDTH)
-        logger.critical(f"Unexpected error: {str(e)}")
-        logger.critical(f"Error type: {type(e).__name__}")
-        logger.exception("Full error details:")
-        logger.critical("")
-        logger.critical("Troubleshooting steps:")
-        logger.critical("1. Verify your configuration file exists and is valid JSON")
-        logger.critical("2. Check that all authentication credentials are correct")
-        logger.critical("3. Ensure your API credentials have the necessary permissions")
-        logger.critical("4. Verify you have network connectivity to Adobe services")
-        logger.critical("5. Check if cjapy library is up to date: pip install --upgrade cjapy")
-        return None
 
 # ==================== DATA VIEW VALIDATION ====================
 
@@ -2885,726 +1055,11 @@ def validate_data_view(
         logger.error("  3. Your API credentials are valid")
         return False
 
-# ==================== OPTIMIZED API DATA FETCHING ====================
-
-class ParallelAPIFetcher:
-    """Fetch multiple API endpoints in parallel using threading.
-
-    Supports optional API auto-tuning and circuit breaker protection.
-
-    Args:
-        cja: CJA API client
-        logger: Logger instance
-        perf_tracker: Performance tracker for timing metrics
-        max_workers: Maximum parallel workers (default: 3)
-        quiet: Suppress progress output (default: False)
-        tuning_config: Optional API worker auto-tuning configuration
-        circuit_breaker: Optional circuit breaker for failure protection
-    """
-
-    def __init__(self, cja: cjapy.CJA, logger: logging.Logger, perf_tracker: 'PerformanceTracker',
-                 max_workers: int = 3, quiet: bool = False,
-                 tuning_config: Optional[APITuningConfig] = None,
-                 circuit_breaker: Optional[CircuitBreaker] = None):
-        self.cja = cja
-        self.logger = logger
-        self.perf_tracker = perf_tracker
-        self.max_workers = max_workers
-        self.quiet = quiet
-        self.circuit_breaker = circuit_breaker
-
-        # Initialize API tuner if config provided
-        self.tuner: Optional[APIWorkerTuner] = None
-        if tuning_config is not None:
-            self.tuner = APIWorkerTuner(
-                config=tuning_config,
-                initial_workers=max_workers,
-                logger=logger
-            )
-            # Use tuner's worker count as effective max
-            self.max_workers = self.tuner.current_workers
-    
-    def fetch_all_data(self, data_view_id: str) -> Tuple[pd.DataFrame, pd.DataFrame, dict]:
-        """
-        Fetch metrics, dimensions, and data view info in parallel
-        
-        Returns:
-            Tuple of (metrics_df, dimensions_df, dataview_info)
-        """
-        self.logger.info("Starting parallel data fetch operations...")
-        self.perf_tracker.start("Parallel API Fetch")
-
-        results = {
-            'metrics': None,
-            'dimensions': None,
-            'dataview': None
-        }
-
-        errors = {}
-
-        # Define fetch tasks
-        tasks = {
-            'metrics': lambda: self._fetch_metrics(data_view_id),
-            'dimensions': lambda: self._fetch_dimensions(data_view_id),
-            'dataview': lambda: self._fetch_dataview_info(data_view_id)
-        }
-
-        # Execute tasks in parallel using ThreadPoolExecutor
-        with ThreadPoolExecutor(max_workers=self.max_workers) as executor:
-            # Submit all tasks
-            future_to_name = {
-                executor.submit(task): name
-                for name, task in tasks.items()
-            }
-
-            # Collect results as they complete with progress indicator
-            with tqdm(
-                total=len(tasks),
-                desc="Fetching API data",
-                unit="item",
-                bar_format=TQDM_BAR_FORMAT,
-                leave=False,
-                disable=self.quiet
-            ) as pbar:
-                for future in as_completed(future_to_name):
-                    task_name = future_to_name[future]
-                    try:
-                        results[task_name] = future.result()
-                        pbar.set_postfix_str(f"✓ {task_name}", refresh=True)
-                        self.logger.info(f"✓ {task_name.capitalize()} fetch completed")
-                    except Exception as e:
-                        errors[task_name] = str(e)
-                        pbar.set_postfix_str(f"✗ {task_name}", refresh=True)
-                        self.logger.error(f"✗ {task_name.capitalize()} fetch failed: {e}")
-                    pbar.update(1)
-
-        self.perf_tracker.end("Parallel API Fetch")
-        
-        # Log summary
-        success_count = sum(1 for v in results.values() if v is not None)
-        self.logger.info(f"Parallel fetch complete: {success_count}/3 successful")
-        
-        if errors:
-            self.logger.warning(f"Errors encountered: {list(errors.keys())}")
-        
-        # Return results with proper None checking for DataFrames
-        metrics_result = results.get('metrics')
-        dimensions_result = results.get('dimensions')
-        dataview_result = results.get('dataview')
-        
-        # Handle DataFrame None checks properly
-        if metrics_result is None or (isinstance(metrics_result, pd.DataFrame) and metrics_result.empty):
-            metrics_result = pd.DataFrame()
-        
-        if dimensions_result is None or (isinstance(dimensions_result, pd.DataFrame) and dimensions_result.empty):
-            dimensions_result = pd.DataFrame()
-        
-        if dataview_result is None:
-            dataview_result = {}
-        
-        return metrics_result, dimensions_result, dataview_result
-    
-    def _timed_api_call(self, api_func: Callable, *args, operation_name: str, **kwargs) -> Any:
-        """
-        Execute an API call with timing for auto-tuning.
-
-        Wraps make_api_call_with_retry with timing measurement and tuner feedback.
-        Only records timing on successful calls so retries don't inflate metrics.
-        """
-        start_time = time.time()
-        result = make_api_call_with_retry(
-            api_func,
-            *args,
-            logger=self.logger,
-            operation_name=operation_name,
-            circuit_breaker=self.circuit_breaker,
-            **kwargs
-        )
-        # Record response time only on success to avoid retry delays inflating metrics
-        if self.tuner is not None:
-            duration_ms = (time.time() - start_time) * 1000
-            new_workers = self.tuner.record_response_time(duration_ms)
-            if new_workers is not None:
-                self.max_workers = new_workers
-        return result
-
-    def _fetch_metrics(self, data_view_id: str) -> pd.DataFrame:
-        """Fetch metrics with error handling and retry"""
-        try:
-            self.logger.debug(f"Fetching metrics for {data_view_id}")
-
-            # Use retry for transient network errors with circuit breaker support
-            metrics = self._timed_api_call(
-                self.cja.getMetrics,
-                data_view_id,
-                inclType=True,
-                full=True,
-                operation_name="getMetrics"
-            )
-
-            if metrics is None or (isinstance(metrics, pd.DataFrame) and metrics.empty):
-                self.logger.warning("No metrics returned from API")
-                return pd.DataFrame()
-
-            self.logger.info(f"Successfully fetched {len(metrics)} metrics")
-            return metrics
-
-        except CircuitBreakerOpen as e:
-            self.logger.warning(f"Circuit breaker open for metrics fetch: {e.message}")
-            return pd.DataFrame()
-        except AttributeError as e:
-            self.logger.error(f"API method error - getMetrics may not be available: {str(e)}")
-            return pd.DataFrame()
-        except Exception as e:
-            self.logger.error(f"Failed to fetch metrics: {str(e)}")
-            return pd.DataFrame()
-
-    def _fetch_dimensions(self, data_view_id: str) -> pd.DataFrame:
-        """Fetch dimensions with error handling and retry"""
-        try:
-            self.logger.debug(f"Fetching dimensions for {data_view_id}")
-
-            # Use retry for transient network errors with circuit breaker support
-            dimensions = self._timed_api_call(
-                self.cja.getDimensions,
-                data_view_id,
-                inclType=True,
-                full=True,
-                operation_name="getDimensions"
-            )
-
-            if dimensions is None or (isinstance(dimensions, pd.DataFrame) and dimensions.empty):
-                self.logger.warning("No dimensions returned from API")
-                return pd.DataFrame()
-
-            self.logger.info(f"Successfully fetched {len(dimensions)} dimensions")
-            return dimensions
-
-        except CircuitBreakerOpen as e:
-            self.logger.warning(f"Circuit breaker open for dimensions fetch: {e.message}")
-            return pd.DataFrame()
-        except AttributeError as e:
-            self.logger.error(f"API method error - getDimensions may not be available: {str(e)}")
-            return pd.DataFrame()
-        except Exception as e:
-            self.logger.error(f"Failed to fetch dimensions: {str(e)}")
-            return pd.DataFrame()
-
-    def _fetch_dataview_info(self, data_view_id: str) -> dict:
-        """Fetch data view information with error handling and retry"""
-        try:
-            self.logger.debug(f"Fetching data view information for {data_view_id}")
-
-            # Use retry for transient network errors with circuit breaker support
-            lookup_data = self._timed_api_call(
-                self.cja.getDataView,
-                data_view_id,
-                operation_name="getDataView"
-            )
-
-            if not lookup_data:
-                self.logger.error("Data view information returned empty")
-                return {"name": "Unknown", "id": data_view_id}
-
-            self.logger.info(f"Successfully fetched data view info: {lookup_data.get('name', 'Unknown')}")
-            return lookup_data
-
-        except CircuitBreakerOpen as e:
-            self.logger.warning(f"Circuit breaker open for data view fetch: {e.message}")
-            return {"name": "Unknown", "id": data_view_id, "circuit_breaker_open": True}
-        except Exception as e:
-            self.logger.error(f"Failed to fetch data view information: {str(e)}")
-            return {"name": "Unknown", "id": data_view_id, "error": str(e)}
-
-    def get_tuner_statistics(self) -> Optional[Dict[str, Any]]:
-        """Get API tuner statistics if tuning is enabled."""
-        if self.tuner is not None:
-            return self.tuner.get_statistics()
-        return None
-
-# ==================== DATA QUALITY VALIDATION ====================
-
-class DataQualityChecker:
-    # Severity levels in priority order (highest to lowest) for proper sorting
-    SEVERITY_ORDER = ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW', 'INFO']
-
-    def __init__(self, logger: logging.Logger, validation_cache: Optional[ValidationCache] = None,
-                 quiet: bool = False):
-        self.issues = []
-        self.logger = logger
-        self.validation_cache = validation_cache  # Optional cache for performance
-        self._issues_lock = threading.Lock()  # Thread safety for parallel validation
-        self.quiet = quiet
-    
-    def add_issue(self, severity: str, category: str, item_type: str,
-                  item_name: str, description: str, details: str = ""):
-        """Add a data quality issue to the tracker (thread-safe)"""
-        issue = {
-            'Severity': severity,
-            'Category': category,
-            'Type': item_type,
-            'Item Name': item_name,
-            'Issue': description,
-            'Details': details
-        }
-
-        # Thread-safe append operation
-        with self._issues_lock:
-            self.issues.append(issue)
-
-        # Conditional logging based on log level for performance
-        # Only log individual issues in DEBUG mode
-        if self.logger.isEnabledFor(logging.DEBUG):
-            self.logger.debug(f"DQ Issue [{severity}] - {item_type}: {description}")
-        elif severity in ['CRITICAL', 'HIGH'] and self.logger.isEnabledFor(logging.WARNING):
-            # In non-DEBUG modes, only log CRITICAL/HIGH severity issues
-            self.logger.warning(f"DQ Issue [{severity}] - {item_type}: {description}")
-    
-    def check_duplicates(self, df: pd.DataFrame, item_type: str):
-        """Check for duplicate names in metrics or dimensions"""
-        try:
-            if df.empty:
-                self.logger.info(f"Skipping duplicate check for empty {item_type} dataframe")
-                return
-            
-            if 'name' not in df.columns:
-                self.logger.warning(f"'name' column not found in {item_type}. Skipping duplicate check.")
-                return
-            
-            duplicates = df['name'].value_counts()
-            duplicates = duplicates[duplicates > 1]
-            
-            for name, count in duplicates.items():
-                self.add_issue(
-                    severity='HIGH',
-                    category='Duplicates',
-                    item_type=item_type,
-                    item_name=str(name),
-                    description=f'Duplicate name found {count} times',
-                    details=f'This {item_type.lower()} name appears {count} times in the data view'
-                )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking duplicates", item_type, e))
-    
-    def check_required_fields(self, df: pd.DataFrame, item_type: str, 
-                            required_fields: List[str]):
-        """Validate that required fields are present"""
-        try:
-            if df.empty:
-                self.logger.info(f"Skipping required fields check for empty {item_type} dataframe")
-                return
-            
-            missing_fields = [field for field in required_fields if field not in df.columns]
-            
-            if missing_fields:
-                self.add_issue(
-                    severity='CRITICAL',
-                    category='Missing Fields',
-                    item_type=item_type,
-                    item_name='N/A',
-                    description=f'Required fields missing from API response',
-                    details=f'Missing fields: {", ".join(missing_fields)}'
-                )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking required fields", item_type, e))
-    
-    def check_null_values(self, df: pd.DataFrame, item_type: str, 
-                         critical_fields: List[str]):
-        """Check for null values in critical fields"""
-        try:
-            if df.empty:
-                self.logger.info(f"Skipping null value check for empty {item_type} dataframe")
-                return
-            
-            for field in critical_fields:
-                if field in df.columns:
-                    null_count = df[field].isna().sum()
-                    if null_count > 0:
-                        null_items = df[df[field].isna()]['name'].tolist() if 'name' in df.columns else []
-                        self.add_issue(
-                            severity='MEDIUM',
-                            category='Null Values',
-                            item_type=item_type,
-                            item_name=', '.join(str(x) for x in null_items),
-                            description=f'Null values in "{field}" field',
-                            details=f'{null_count} item(s) missing {field}. Items: {", ".join(str(x) for x in null_items)}'
-                        )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking null values", item_type, e))
-    
-    def check_missing_descriptions(self, df: pd.DataFrame, item_type: str):
-        """Check for items without descriptions"""
-        try:
-            if df.empty:
-                self.logger.info(f"Skipping description check for empty {item_type} dataframe")
-                return
-            
-            if 'description' not in df.columns:
-                self.logger.info(f"'description' column not found in {item_type}")
-                return
-            
-            missing_desc = df[df['description'].isna() | (df['description'] == '')]
-            
-            if len(missing_desc) > 0:
-                item_names = missing_desc['name'].tolist() if 'name' in missing_desc.columns else []
-                self.add_issue(
-                    severity='LOW',
-                    category='Missing Descriptions',
-                    item_type=item_type,
-                    item_name=f'{len(missing_desc)} items',
-                    description=f'{len(missing_desc)} items without descriptions',
-                    details=f'Items: {", ".join(str(x) for x in item_names)}'
-                )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking descriptions", item_type, e))
-    
-    def check_empty_dataframe(self, df: pd.DataFrame, item_type: str):
-        """Check if dataframe is empty"""
-        try:
-            if df.empty:
-                self.add_issue(
-                    severity='CRITICAL',
-                    category='Empty Data',
-                    item_type=item_type,
-                    item_name='N/A',
-                    description=f'No {item_type.lower()} found in data view',
-                    details=f'The API returned an empty dataset for {item_type.lower()}'
-                )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking if dataframe is empty", item_type, e))
-    
-    def check_id_validity(self, df: pd.DataFrame, item_type: str):
-        """Check for missing or invalid IDs"""
-        try:
-            if df.empty:
-                self.logger.info(f"Skipping ID validity check for empty {item_type} dataframe")
-                return
-            
-            if 'id' not in df.columns:
-                self.logger.warning(f"'id' column not found in {item_type}")
-                return
-            
-            missing_ids = df[df['id'].isna() | (df['id'] == '')]
-            if len(missing_ids) > 0:
-                self.add_issue(
-                    severity='HIGH',
-                    category='Invalid IDs',
-                    item_type=item_type,
-                    item_name=f'{len(missing_ids)} items',
-                    description=f'{len(missing_ids)} items with missing or invalid IDs',
-                    details='Items without valid IDs may cause issues in reporting'
-                )
-        except Exception as e:
-            self.logger.error(_format_error_msg("checking ID validity", item_type, e))
-    
-    def check_all_quality_issues_optimized(self, df: pd.DataFrame, item_type: str,
-                                           required_fields: List[str],
-                                           critical_fields: List[str]):
-        """
-        Optimized single-pass validation combining all checks
-
-        PERFORMANCE OPTIMIZATIONS:
-        - 40-55% faster than sequential individual checks
-        - Reduces DataFrame scans from 6 to 1
-        - Uses vectorized pandas operations
-        - Early exit on critical errors (5-10% additional improvement)
-        - Validation caching (50-90% improvement on cache hits)
-        - Better CPU cache utilization
-
-        Args:
-            df: DataFrame to validate (metrics or dimensions)
-            item_type: Type of items ('Metrics' or 'Dimensions')
-            required_fields: Fields that must be present in the DataFrame
-            critical_fields: Fields to check for null values
-
-        Early Exit Conditions:
-            - Cache hit: Returns cached results immediately
-            - Empty DataFrame: Exits immediately
-            - Missing required fields: Exits after logging critical error
-        """
-        try:
-            # Check cache first (before any processing)
-            # get() returns (issues, cache_key) - reuse cache_key in put() to avoid rehashing
-            cache_key = None
-            if self.validation_cache is not None:
-                cached_issues, cache_key = self.validation_cache.get(
-                    df, item_type, required_fields, critical_fields
-                )
-                if cached_issues is not None:
-                    # Cache hit - add issues to tracker and return
-                    with self._issues_lock:
-                        self.issues.extend(cached_issues)
-                    self.logger.debug(f"Using cached validation results for {item_type}")
-                    return
-
-            # Track where validation starts (for caching results at the end)
-            issues_start_index = len(self.issues)
-
-            # Check 1: Empty DataFrame (quick exit)
-            if df.empty:
-                self.add_issue(
-                    severity='CRITICAL',
-                    category='Empty Data',
-                    item_type=item_type,
-                    item_name='N/A',
-                    description=f'No {item_type.lower()} found in data view',
-                    details=f'The API returned an empty dataset for {item_type.lower()}'
-                )
-                # Cache the result before returning (reuse cache_key to avoid rehashing)
-                if self.validation_cache is not None:
-                    new_issues = self.issues[issues_start_index:]
-                    self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
-                return
-
-            # Check 2: Required fields validation (no iteration needed)
-            missing_fields = [field for field in required_fields if field not in df.columns]
-            if missing_fields:
-                self.add_issue(
-                    severity='CRITICAL',
-                    category='Missing Fields',
-                    item_type=item_type,
-                    item_name='N/A',
-                    description='Required fields missing from API response',
-                    details=f'Missing fields: {", ".join(missing_fields)}'
-                )
-                # Cache the critical error result before returning (reuse cache_key)
-                if self.validation_cache is not None:
-                    new_issues = self.issues[issues_start_index:]
-                    self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
-                return  # Early exit: cannot proceed without required fields
-
-            # Check 3: Vectorized duplicate detection
-            if 'name' in df.columns:
-                duplicates = df['name'].value_counts()
-                duplicates = duplicates[duplicates > 1]
-
-                for name, count in duplicates.items():
-                    self.add_issue(
-                        severity='HIGH',
-                        category='Duplicates',
-                        item_type=item_type,
-                        item_name=str(name),
-                        description=f'Duplicate name found {count} times',
-                        details=f'This {item_type.lower()} name appears {count} times in the data view'
-                    )
-
-            # Check 4: Vectorized null value checks (single operation for all fields)
-            available_critical_fields = [f for f in critical_fields if f in df.columns]
-            if available_critical_fields:
-                # Single vectorized operation instead of looping
-                null_counts = df[available_critical_fields].isna().sum()
-
-                for field, null_count in null_counts[null_counts > 0].items():
-                    null_items = df[df[field].isna()]['name'].tolist() if 'name' in df.columns else []
-                    self.add_issue(
-                        severity='MEDIUM',
-                        category='Null Values',
-                        item_type=item_type,
-                        item_name=', '.join(str(x) for x in null_items),
-                        description=f'Null values in "{field}" field',
-                        details=f'{null_count} item(s) missing {field}. Items: {", ".join(str(x) for x in null_items)}'
-                    )
-
-            # Check 5: Vectorized missing descriptions check
-            if 'description' in df.columns:
-                missing_desc = df[df['description'].isna() | (df['description'] == '')]
-
-                if len(missing_desc) > 0:
-                    item_names = missing_desc['name'].tolist() if 'name' in missing_desc.columns else []
-                    self.add_issue(
-                        severity='LOW',
-                        category='Missing Descriptions',
-                        item_type=item_type,
-                        item_name=f'{len(missing_desc)} items',
-                        description=f'{len(missing_desc)} items without descriptions',
-                        details=f'Items: {", ".join(str(x) for x in item_names)}'
-                    )
-
-            # Check 6: Vectorized ID validity check
-            if 'id' in df.columns:
-                missing_ids = df[df['id'].isna() | (df['id'] == '')]
-
-                if len(missing_ids) > 0:
-                    self.add_issue(
-                        severity='HIGH',
-                        category='Invalid IDs',
-                        item_type=item_type,
-                        item_name=f'{len(missing_ids)} items',
-                        description=f'{len(missing_ids)} items with missing or invalid IDs',
-                        details='Items without valid IDs may cause issues in reporting'
-                    )
-
-            if self.logger.isEnabledFor(logging.DEBUG):
-                self.logger.debug(f"Optimized validation complete for {item_type}: {len(df)} items checked")
-
-            # Store results in cache after successful validation (reuse cache_key to avoid rehashing)
-            if self.validation_cache is not None:
-                new_issues = self.issues[issues_start_index:]
-                self.validation_cache.put(df, item_type, required_fields, critical_fields, new_issues, cache_key)
-
-        except Exception as e:
-            self.logger.error(_format_error_msg("in optimized validation", item_type, e))
-            self.logger.exception("Full error details:")
-
-    def check_all_parallel(self,
-                          metrics_df: pd.DataFrame,
-                          dimensions_df: pd.DataFrame,
-                          metrics_required_fields: List[str],
-                          dimensions_required_fields: List[str],
-                          critical_fields: List[str],
-                          max_workers: int = 2):
-        """
-        Run validation checks in parallel for metrics and dimensions
-
-        PERFORMANCE: 10-15% faster than sequential validation
-        - Uses ThreadPoolExecutor to validate metrics and dimensions concurrently
-        - Thread-safe issue collection using locks
-        - Maintains identical validation results to sequential method
-
-        Args:
-            metrics_df: DataFrame containing metrics data
-            dimensions_df: DataFrame containing dimensions data
-            metrics_required_fields: Required fields for metrics validation
-            dimensions_required_fields: Required fields for dimensions validation
-            critical_fields: Fields to check for null values (shared across both)
-            max_workers: Number of worker threads (default: 2, one for metrics, one for dimensions)
-
-        Returns:
-            None (issues are stored in self.issues)
-
-        Thread Safety:
-            - Uses self._issues_lock to protect shared self.issues list
-            - Each validation task runs independently on separate DataFrames
-            - Logging module is inherently thread-safe
-        """
-        try:
-            self.logger.info("Starting parallel validation (metrics and dimensions)")
-
-            # Define validation tasks
-            tasks = {
-                'metrics': lambda: self.check_all_quality_issues_optimized(
-                    metrics_df, 'Metrics', metrics_required_fields, critical_fields
-                ),
-                'dimensions': lambda: self.check_all_quality_issues_optimized(
-                    dimensions_df, 'Dimensions', dimensions_required_fields, critical_fields
-                )
-            }
-
-            # Execute validations in parallel
-            with ThreadPoolExecutor(max_workers=max_workers) as executor:
-                # Submit all validation tasks
-                future_to_name = {
-                    executor.submit(task): name
-                    for name, task in tasks.items()
-                }
-
-                # Collect results as they complete with progress indicator
-                with tqdm(
-                    total=len(tasks),
-                    desc="Validating data",
-                    unit="check",
-                    bar_format=TQDM_BAR_FORMAT,
-                    leave=False,
-                    disable=self.quiet
-                ) as pbar:
-                    for future in as_completed(future_to_name):
-                        task_name = future_to_name[future]
-                        try:
-                            future.result()  # This will re-raise any exception from the task
-                            pbar.set_postfix_str(f"✓ {task_name}", refresh=True)
-                            self.logger.debug(f"✓ {task_name.capitalize()} validation completed")
-                        except Exception as e:
-                            pbar.set_postfix_str(f"✗ {task_name}", refresh=True)
-                            self.logger.error(f"✗ {task_name.capitalize()} validation failed: {e}")
-                            self.logger.exception("Full error details:")
-                        pbar.update(1)
-
-            self.logger.info(f"Parallel validation complete. Found {len(self.issues)} issue(s)")
-
-        except Exception as e:
-            self.logger.error(_format_error_msg("in parallel validation", error=e))
-            self.logger.exception("Full error details:")
-            raise
-
-    def get_issues_dataframe(self, max_issues: int = 0) -> pd.DataFrame:
-        """Return all issues as a DataFrame sorted by severity (CRITICAL first)
-
-        Args:
-            max_issues: Maximum number of issues to return (0 = all issues)
-        """
-        try:
-            if not self.issues:
-                self.logger.info("No data quality issues found")
-                return pd.DataFrame({
-                    'Severity': ['INFO'],
-                    'Category': ['Data Quality'],
-                    'Type': ['All'],
-                    'Item Name': ['N/A'],
-                    'Issue': ['No data quality issues detected'],
-                    'Details': ['All validation checks passed successfully']
-                })
-
-            df = pd.DataFrame(self.issues)
-
-            # Use CategoricalDtype for proper severity ordering (CRITICAL > HIGH > MEDIUM > LOW > INFO)
-            severity_dtype = pd.CategoricalDtype(categories=self.SEVERITY_ORDER, ordered=True)
-            df['Severity'] = df['Severity'].astype(severity_dtype)
-
-            # Reorder columns: Severity first for better readability
-            preferred_order = ['Severity', 'Category', 'Type', 'Item Name', 'Issue', 'Details']
-            existing_cols = [col for col in preferred_order if col in df.columns]
-            other_cols = [col for col in df.columns if col not in preferred_order]
-            df = df[existing_cols + other_cols]
-
-            # Sort by severity (ascending=True with ordered categorical puts CRITICAL first)
-            # then by Category alphabetically
-            df = df.sort_values(
-                by=['Severity', 'Category'],
-                ascending=[True, True]
-            )
-
-            # Limit to top N issues if max_issues > 0
-            if max_issues > 0 and len(df) > max_issues:
-                self.logger.info(f"Limiting data quality issues to top {max_issues} (of {len(df)} total)")
-                df = df.head(max_issues)
-
-            return df
-        except Exception as e:
-            self.logger.error(_format_error_msg("creating issues dataframe", error=e))
-            return pd.DataFrame({
-                'Severity': ['ERROR'],
-                'Category': ['System'],
-                'Type': ['Processing'],
-                'Item Name': ['N/A'],
-                'Issue': ['Error generating data quality report'],
-                'Details': [str(e)]
-            })
-
-    def log_summary(self):
-        """Log aggregated summary of data quality issues for performance
-
-        Instead of logging each individual issue (which can be 100+ log entries),
-        this method logs a concise summary with counts by severity.
-        Individual issue details are still captured in self.issues list.
-        """
-        if not self.issues:
-            self.logger.info("✓ No data quality issues found")
-            return
-
-        # Aggregate by severity
-        severity_counts = {}
-        for issue in self.issues:
-            sev = issue['Severity']
-            severity_counts[sev] = severity_counts.get(sev, 0) + 1
-
-        # Log summary
-        self.logger.info(f"Data quality validation complete: {len(self.issues)} issue(s) found")
-
-        # Log severity breakdown at INFO level
-        if self.logger.isEnabledFor(logging.INFO):
-            for sev in ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW']:
-                if sev in severity_counts:
-                    self.logger.info(f"  {sev}: {severity_counts[sev]}")
+# ==================== OPTIMIZED API DATA FETCHING (moved to api/fetch.py) ====================
+from cja_auto_sdr.api.fetch import ParallelAPIFetcher  # noqa: F811
+
+# ==================== DATA QUALITY VALIDATION (moved to api/quality.py) ====================
+from cja_auto_sdr.api.quality import DataQualityChecker  # noqa: F811
 
 # ==================== EXCEL GENERATION ====================
